Access or assign mass at age

This function can be used to access or assign the biomass at age in an lhm object.

Assignment can be of a list or numeric vector. When a list is assigned it should provide the \(a\) and \(b\) parameters of the function: $$Mass = a * Length ^ b.$$

Uncertainty accross iterations in the lhm object can be included by providing either an overall coefficient of variation (cv) or a cv for each of the parameters. Overall uncertainty is assumed to be log-normal with \(\sigma = \sqrt{ln(1+cv^2)}\). Parameter uncertainty is assumed to be normal with \(\sigma = \mu * cv\).

When assigning a numeric vector it is assumed that each replicate is identical.

Usage

mass(object, ...)

# S4 method for class 'lhm'
mass(object)

mass(object) <- value

# S4 method for class 'lhm,list'
mass(object) <- value

# S4 method for class 'lhm,numeric'
mass(object) <- value

Arguments

object

a lhm object

...

arguments to generic function

value

either a named list or numeric vector

numeric

a vector of length equal to the maximum age

list

a list object containing the components mu and cv

mu

a list containing the parameters a and b representing their mean values

cv

optional single numeric value or a list containing parameters a and b representing their coefficients of variation

Value

Accessor function returns a matrix of mass at age across iterations. Assignment function populates the matrix across iterations.

Examples

# initialize single iteration
dat <- lhm(ainf = 30, iter = 1)

# add length data
size(dat) <- list(mu=list(Linf = 100, k = 0.1, t0 = -0.5))
size(dat)
#>           [,1]
#>  [1,] 13.92920
#>  [2,] 22.11992
#>  [3,] 29.53119
#>  [4,] 36.23718
#>  [5,] 42.30502
#>  [6,] 47.79542
#>  [7,] 52.76334
#>  [8,] 57.25851
#>  [9,] 61.32590
#> [10,] 65.00623
#> [11,] 68.33632
#> [12,] 71.34952
#> [13,] 74.07597
#> [14,] 76.54297
#> [15,] 78.77520
#> [16,] 80.79501
#> [17,] 82.62261
#> [18,] 84.27628
#> [19,] 85.77259
#> [20,] 87.12651
#> [21,] 88.35158
#> [22,] 89.46008
#> [23,] 90.46308
#> [24,] 91.37064
#> [25,] 92.19183
#> [26,] 92.93488
#> [27,] 93.60721
#> [28,] 94.21557
#> [29,] 94.76603
#> [30,] 95.26411

# mass at age
mass(dat) <- list(mu = list(a = 9.63e-6, b = 3.173))
mass(dat)
#>              [,1]
#>  [1,]  0.04104916
#>  [2,]  0.17808388
#>  [3,]  0.44547956
#>  [4,]  0.85275321
#>  [5,]  1.39369508
#>  [6,]  2.05266949
#>  [7,]  2.80922635
#>  [8,]  3.64126817
#>  [9,]  4.52710352
#> [10,]  5.44668595
#> [11,]  6.38227976
#> [12,]  7.31873653
#> [13,]  8.24351785
#> [14,]  9.14656144
#> [15,] 10.02005862
#> [16,] 10.85818945
#> [17,] 11.65684592
#> [18,] 12.41336259
#> [19,] 13.12626598
#> [20,] 13.79504878
#> [21,] 14.41997116
#> [22,] 15.00188948
#> [23,] 15.54211085
#> [24,] 16.04227148
#> [25,] 16.50423644
#> [26,] 16.93001792
#> [27,] 17.32170973
#> [28,] 17.68143546
#> [29,] 18.01130809
#> [30,] 18.31339938

# iterate
dat <- iterate(dat, iter = 10)
size(dat)
#>           [,1]     [,2]     [,3]     [,4]     [,5]     [,6]     [,7]     [,8]
#>  [1,] 13.92920 13.92920 13.92920 13.92920 13.92920 13.92920 13.92920 13.92920
#>  [2,] 22.11992 22.11992 22.11992 22.11992 22.11992 22.11992 22.11992 22.11992
#>  [3,] 29.53119 29.53119 29.53119 29.53119 29.53119 29.53119 29.53119 29.53119
#>  [4,] 36.23718 36.23718 36.23718 36.23718 36.23718 36.23718 36.23718 36.23718
#>  [5,] 42.30502 42.30502 42.30502 42.30502 42.30502 42.30502 42.30502 42.30502
#>  [6,] 47.79542 47.79542 47.79542 47.79542 47.79542 47.79542 47.79542 47.79542
#>  [7,] 52.76334 52.76334 52.76334 52.76334 52.76334 52.76334 52.76334 52.76334
#>  [8,] 57.25851 57.25851 57.25851 57.25851 57.25851 57.25851 57.25851 57.25851
#>  [9,] 61.32590 61.32590 61.32590 61.32590 61.32590 61.32590 61.32590 61.32590
#> [10,] 65.00623 65.00623 65.00623 65.00623 65.00623 65.00623 65.00623 65.00623
#> [11,] 68.33632 68.33632 68.33632 68.33632 68.33632 68.33632 68.33632 68.33632
#> [12,] 71.34952 71.34952 71.34952 71.34952 71.34952 71.34952 71.34952 71.34952
#> [13,] 74.07597 74.07597 74.07597 74.07597 74.07597 74.07597 74.07597 74.07597
#> [14,] 76.54297 76.54297 76.54297 76.54297 76.54297 76.54297 76.54297 76.54297
#> [15,] 78.77520 78.77520 78.77520 78.77520 78.77520 78.77520 78.77520 78.77520
#> [16,] 80.79501 80.79501 80.79501 80.79501 80.79501 80.79501 80.79501 80.79501
#> [17,] 82.62261 82.62261 82.62261 82.62261 82.62261 82.62261 82.62261 82.62261
#> [18,] 84.27628 84.27628 84.27628 84.27628 84.27628 84.27628 84.27628 84.27628
#> [19,] 85.77259 85.77259 85.77259 85.77259 85.77259 85.77259 85.77259 85.77259
#> [20,] 87.12651 87.12651 87.12651 87.12651 87.12651 87.12651 87.12651 87.12651
#> [21,] 88.35158 88.35158 88.35158 88.35158 88.35158 88.35158 88.35158 88.35158
#> [22,] 89.46008 89.46008 89.46008 89.46008 89.46008 89.46008 89.46008 89.46008
#> [23,] 90.46308 90.46308 90.46308 90.46308 90.46308 90.46308 90.46308 90.46308
#> [24,] 91.37064 91.37064 91.37064 91.37064 91.37064 91.37064 91.37064 91.37064
#> [25,] 92.19183 92.19183 92.19183 92.19183 92.19183 92.19183 92.19183 92.19183
#> [26,] 92.93488 92.93488 92.93488 92.93488 92.93488 92.93488 92.93488 92.93488
#> [27,] 93.60721 93.60721 93.60721 93.60721 93.60721 93.60721 93.60721 93.60721
#> [28,] 94.21557 94.21557 94.21557 94.21557 94.21557 94.21557 94.21557 94.21557
#> [29,] 94.76603 94.76603 94.76603 94.76603 94.76603 94.76603 94.76603 94.76603
#> [30,] 95.26411 95.26411 95.26411 95.26411 95.26411 95.26411 95.26411 95.26411
#>           [,9]    [,10]
#>  [1,] 13.92920 13.92920
#>  [2,] 22.11992 22.11992
#>  [3,] 29.53119 29.53119
#>  [4,] 36.23718 36.23718
#>  [5,] 42.30502 42.30502
#>  [6,] 47.79542 47.79542
#>  [7,] 52.76334 52.76334
#>  [8,] 57.25851 57.25851
#>  [9,] 61.32590 61.32590
#> [10,] 65.00623 65.00623
#> [11,] 68.33632 68.33632
#> [12,] 71.34952 71.34952
#> [13,] 74.07597 74.07597
#> [14,] 76.54297 76.54297
#> [15,] 78.77520 78.77520
#> [16,] 80.79501 80.79501
#> [17,] 82.62261 82.62261
#> [18,] 84.27628 84.27628
#> [19,] 85.77259 85.77259
#> [20,] 87.12651 87.12651
#> [21,] 88.35158 88.35158
#> [22,] 89.46008 89.46008
#> [23,] 90.46308 90.46308
#> [24,] 91.37064 91.37064
#> [25,] 92.19183 92.19183
#> [26,] 92.93488 92.93488
#> [27,] 93.60721 93.60721
#> [28,] 94.21557 94.21557
#> [29,] 94.76603 94.76603
#> [30,] 95.26411 95.26411
mass(dat)
#>              [,1]        [,2]        [,3]        [,4]        [,5]        [,6]
#>  [1,]  0.04104916  0.04104916  0.04104916  0.04104916  0.04104916  0.04104916
#>  [2,]  0.17808388  0.17808388  0.17808388  0.17808388  0.17808388  0.17808388
#>  [3,]  0.44547956  0.44547956  0.44547956  0.44547956  0.44547956  0.44547956
#>  [4,]  0.85275321  0.85275321  0.85275321  0.85275321  0.85275321  0.85275321
#>  [5,]  1.39369508  1.39369508  1.39369508  1.39369508  1.39369508  1.39369508
#>  [6,]  2.05266949  2.05266949  2.05266949  2.05266949  2.05266949  2.05266949
#>  [7,]  2.80922635  2.80922635  2.80922635  2.80922635  2.80922635  2.80922635
#>  [8,]  3.64126817  3.64126817  3.64126817  3.64126817  3.64126817  3.64126817
#>  [9,]  4.52710352  4.52710352  4.52710352  4.52710352  4.52710352  4.52710352
#> [10,]  5.44668595  5.44668595  5.44668595  5.44668595  5.44668595  5.44668595
#> [11,]  6.38227976  6.38227976  6.38227976  6.38227976  6.38227976  6.38227976
#> [12,]  7.31873653  7.31873653  7.31873653  7.31873653  7.31873653  7.31873653
#> [13,]  8.24351785  8.24351785  8.24351785  8.24351785  8.24351785  8.24351785
#> [14,]  9.14656144  9.14656144  9.14656144  9.14656144  9.14656144  9.14656144
#> [15,] 10.02005862 10.02005862 10.02005862 10.02005862 10.02005862 10.02005862
#> [16,] 10.85818945 10.85818945 10.85818945 10.85818945 10.85818945 10.85818945
#> [17,] 11.65684592 11.65684592 11.65684592 11.65684592 11.65684592 11.65684592
#> [18,] 12.41336259 12.41336259 12.41336259 12.41336259 12.41336259 12.41336259
#> [19,] 13.12626598 13.12626598 13.12626598 13.12626598 13.12626598 13.12626598
#> [20,] 13.79504878 13.79504878 13.79504878 13.79504878 13.79504878 13.79504878
#> [21,] 14.41997116 14.41997116 14.41997116 14.41997116 14.41997116 14.41997116
#> [22,] 15.00188948 15.00188948 15.00188948 15.00188948 15.00188948 15.00188948
#> [23,] 15.54211085 15.54211085 15.54211085 15.54211085 15.54211085 15.54211085
#> [24,] 16.04227148 16.04227148 16.04227148 16.04227148 16.04227148 16.04227148
#> [25,] 16.50423644 16.50423644 16.50423644 16.50423644 16.50423644 16.50423644
#> [26,] 16.93001792 16.93001792 16.93001792 16.93001792 16.93001792 16.93001792
#> [27,] 17.32170973 17.32170973 17.32170973 17.32170973 17.32170973 17.32170973
#> [28,] 17.68143546 17.68143546 17.68143546 17.68143546 17.68143546 17.68143546
#> [29,] 18.01130809 18.01130809 18.01130809 18.01130809 18.01130809 18.01130809
#> [30,] 18.31339938 18.31339938 18.31339938 18.31339938 18.31339938 18.31339938
#>              [,7]        [,8]        [,9]       [,10]
#>  [1,]  0.04104916  0.04104916  0.04104916  0.04104916
#>  [2,]  0.17808388  0.17808388  0.17808388  0.17808388
#>  [3,]  0.44547956  0.44547956  0.44547956  0.44547956
#>  [4,]  0.85275321  0.85275321  0.85275321  0.85275321
#>  [5,]  1.39369508  1.39369508  1.39369508  1.39369508
#>  [6,]  2.05266949  2.05266949  2.05266949  2.05266949
#>  [7,]  2.80922635  2.80922635  2.80922635  2.80922635
#>  [8,]  3.64126817  3.64126817  3.64126817  3.64126817
#>  [9,]  4.52710352  4.52710352  4.52710352  4.52710352
#> [10,]  5.44668595  5.44668595  5.44668595  5.44668595
#> [11,]  6.38227976  6.38227976  6.38227976  6.38227976
#> [12,]  7.31873653  7.31873653  7.31873653  7.31873653
#> [13,]  8.24351785  8.24351785  8.24351785  8.24351785
#> [14,]  9.14656144  9.14656144  9.14656144  9.14656144
#> [15,] 10.02005862 10.02005862 10.02005862 10.02005862
#> [16,] 10.85818945 10.85818945 10.85818945 10.85818945
#> [17,] 11.65684592 11.65684592 11.65684592 11.65684592
#> [18,] 12.41336259 12.41336259 12.41336259 12.41336259
#> [19,] 13.12626598 13.12626598 13.12626598 13.12626598
#> [20,] 13.79504878 13.79504878 13.79504878 13.79504878
#> [21,] 14.41997116 14.41997116 14.41997116 14.41997116
#> [22,] 15.00188948 15.00188948 15.00188948 15.00188948
#> [23,] 15.54211085 15.54211085 15.54211085 15.54211085
#> [24,] 16.04227148 16.04227148 16.04227148 16.04227148
#> [25,] 16.50423644 16.50423644 16.50423644 16.50423644
#> [26,] 16.93001792 16.93001792 16.93001792 16.93001792
#> [27,] 17.32170973 17.32170973 17.32170973 17.32170973
#> [28,] 17.68143546 17.68143546 17.68143546 17.68143546
#> [29,] 18.01130809 18.01130809 18.01130809 18.01130809
#> [30,] 18.31339938 18.31339938 18.31339938 18.31339938

# multiple stochastic iterations
# with overall uncertainty
mass(dat) <- list(mu = list(a = 9.63e-6,b = 3.173), cv = 0.05)
mass(dat)
#>              [,1]       [,2]        [,3]        [,4]       [,5]        [,6]
#>  [1,]  0.02940388  0.0423840  0.02338326  0.04001061  0.0459552  0.05162686
#>  [2,]  0.12756309  0.1838748  0.10144378  0.17357833  0.1993678  0.22397321
#>  [3,]  0.31910102  0.4599657  0.25376317  0.43420886  0.4987216  0.56027242
#>  [4,]  0.61083480  0.8804832  0.48576270  0.83117843  0.9546710  1.07249388
#>  [5,]  0.99831634  1.4390155  0.79390506  1.35843439  1.5602641  1.75282771
#>  [6,]  1.47034564  2.1194185  1.16928352  2.00073666  2.2979966  2.58160914
#>  [7,]  2.01227413  2.9005772  1.60024889  2.73815251  3.1449742  3.53311843
#>  [8,]  2.60827318  3.7596754  2.07421355  3.54914356  4.0764584  4.57956395
#>  [9,]  3.24280502  4.6743165  2.57882118  4.41256715  5.0681654  5.69366471
#> [10,]  3.90151019  5.6238020  3.10265250  5.30888401  6.0976528  6.85020862
#> [11,]  4.57168446  6.5898196  3.63560456  6.22080716  7.1450651  8.02688978
#> [12,]  5.24247688  7.5567282  4.16904819  7.13357144  8.1934436  9.20465628
#> [13,]  5.90490607  8.5115817  4.69584102  8.03495566  9.2287512 10.36773876
#> [14,]  6.55176432  9.4439906  5.21025115  8.91515212 10.2397231 11.50348204
#> [15,]  7.17745822 10.3458923  5.70783046  9.76654969 11.2176172 12.60206528
#> [16,]  7.77781888 11.2112776  6.18526366 10.58347569 12.1559182 13.65616884
#> [17,]  8.34990370 12.0359049  6.64021067 11.36192604 13.0500270 14.66062613
#> [18,]  8.89180341 12.8170220  7.07115315 12.09930273 13.8969596 15.61208489
#> [19,]  9.40246253 13.5531077  7.47725174 12.79416956 14.6950665 16.50869193
#> [20,]  9.88151766 14.2436381  7.85821745 13.44603205 15.4437796 17.34980920
#> [21,] 10.32915519 14.8888818  8.21419851 14.05514381 16.1433903 18.13576395
#> [22,] 10.74598853 15.4897230  8.54568271 14.62233952 16.7948572 18.86763319
#> [23,] 11.13295396 16.0475114  8.85341464 15.14889320 17.3996438 19.54706085
#> [24,] 11.49122352 16.5639363  9.13832635 15.63639970 17.9595817 20.17610477
#> [25,] 11.82213318 17.0409235  9.40148027 16.08667688 18.4767589 20.75711060
#> [26,] 12.12712429 17.4805506  9.64402261 16.50168602 18.9534281 21.29260907
#> [27,] 12.40769666 17.8849795  9.86714610 16.88346798 19.3919334 21.78523352
#> [28,] 12.66537144 18.2564028 10.07206042 17.23409259 19.7946522 22.23765472
#> [29,] 12.90166218 18.5970023 10.25996921 17.55561940 20.1639499 22.65253019
#> [30,] 13.11805289 18.9089171 10.43205261 17.85006773 20.5021460 23.03246550
#>              [,7]        [,8]        [,9]       [,10]
#>  [1,]  0.02678824  0.03793009  0.03795629  0.03763483
#>  [2,]  0.11621564  0.16455241  0.16466606  0.16327146
#>  [3,]  0.29071522  0.41163039  0.41191469  0.40842606
#>  [4,]  0.55649767  0.78795790  0.78850211  0.78182405
#>  [5,]  0.90951058  1.28779703  1.28868646  1.27777219
#>  [6,]  1.33955027  1.89670016  1.89801013  1.88193532
#>  [7,]  1.83327123  2.59577105  2.59756384  2.57556431
#>  [8,]  2.37625287  3.36459130  3.36691508  3.33839969
#>  [9,]  2.95433960  4.18311764  4.18600674  4.15055422
#> [10,]  3.55444930  5.03282683  5.03630279  4.99364886
#> [11,]  4.16500787  5.89733080  5.90140383  5.85142312
#> [12,]  4.77612960  6.76263216  6.76730282  6.70998855
#> [13,]  5.37963205  7.61714522  7.62240605  7.55784966
#> [14,]  5.96894868  8.45157225  8.45740939  8.38578111
#> [15,]  6.53898365  9.25869792  9.26509251  9.18662372
#> [16,]  7.08593891 10.03314450 10.04007397  9.95504164
#> [17,]  7.60713363 10.77111613 10.77855528 10.68726854
#> [18,]  8.10082836 11.47014990 11.47807184 11.38086069
#> [19,]  8.56606153 12.12888429 12.13726119 12.03446717
#> [20,]  9.00250207 12.74685052 12.75565422 12.64762285
#> [21,]  9.41031976 13.32428901 13.33349153 13.22056628
#> [22,]  9.79007347 13.86199105 13.87156494 13.75408259
#> [23,] 10.14261618 14.36116442 14.37108306 14.24937015
#> [24,] 10.46901569 14.82332102 14.83355886 14.70792910
#> [25,] 10.77048911 15.25018419 15.26071684 15.13146936
#> [26,] 11.04834957 15.64361323 15.65441761 15.52183576
#> [27,] 11.30396348 16.00554287 16.01659722 15.88094797
#> [28,] 11.53871666 16.33793532 16.34921924 16.21075291
#> [29,] 11.75398804 16.64274303 16.65423746 16.51318785
#> [30,] 11.95112960 16.92188031 16.93356753 16.79015219

# multiple stochastic iterations
# with parameter uncertainty
mass(dat) <- list(mu = list(a = 9.63e-6,b = 3.173), cv = list(a = 0.05, b = 0.05))
mass(dat)
#>              [,1]        [,2]       [,3]        [,4]        [,5]        [,6]
#>  [1,]  0.04853709  0.04927446 0.02625579  0.05131621  0.09268822  0.03574002
#>  [2,]  0.21792764  0.21953715 0.10350820  0.23500921  0.46187454  0.15394652
#>  [3,]  0.55697605  0.55838999 0.24389578  0.60810323  1.25987551  0.38338305
#>  [4,]  1.08251262  1.08156029 0.44751344  1.19227208  2.56419303  0.73156772
#>  [5,]  1.78966309  1.78347415 0.70832837  1.98421990  4.38975437  1.19277760
#>  [6,]  2.65985879  2.64526762 1.01722697  2.96444886  6.70608733  1.75344222
#>  [7,]  3.66704299  3.64091255 1.36394612  4.10429362  9.45374800  2.39604645
#>  [8,]  4.78210587  4.74155499 1.73825812  5.37106573 12.55765254  3.10178883
#>  [9,]  5.97586205  5.91840563 2.13065748  6.73157550 15.93697191  3.85229228
#> [10,]  7.22091763  7.14453720 2.53271912  8.15438555 19.51197307  4.63062742
#> [11,]  8.49273089  8.39589397 2.93724426  9.61112662 23.20840286  5.42185524
#> [12,]  9.77011034  9.65175481 3.33827350 11.07714892 26.96001206  6.21324514
#> [13,] 11.03533614 10.89483237 3.73102083 12.53172187 30.70973827  6.99428263
#> [14,] 12.27404150 12.11114217 4.11176465 13.95794158 34.40996587  7.75654813
#> [15,] 13.47495210 13.28973672 4.47771914 15.34246134 38.02218402  8.49352357
#> [16,] 14.62955147 14.42237045 4.82690122 16.67512645 41.51628072  9.20036526
#> [17,] 15.73171839 15.50313974 5.15800208 17.94856882 44.86964268  9.87366823
#> [18,] 16.77736589 16.52812648 5.47026857 19.15779774 48.06617875 10.51123781
#> [19,] 17.76410017 17.49506245 5.76339698 20.29980947 51.09534486 11.11187789
#> [20,] 18.69090973 18.40302426 6.03744016 21.37322893 53.95121930 11.67520049
#> [21,] 19.55788926 19.25216300 6.29272778 22.37798984 56.63165662 12.20145860
#> [22,] 20.36599975 20.04346968 6.52979895 23.31505531 59.13753390 12.69140220
#> [23,] 21.11686345 20.77857517 6.74934607 24.18617838 61.47209347 13.14615627
#> [24,] 21.81259142 21.45958201 6.95216867 24.99369976 63.64037991 13.56711880
#> [25,] 22.45564050 22.08892524 7.13913595 25.74037958 65.64876520 13.95587690
#> [26,] 23.04869620 22.66925848 7.31115668 26.42925927 67.50455414 14.31413848
#> [27,] 23.59457829 23.20336220 7.46915560 27.06354975 69.21566090 14.64367749
#> [28,] 24.09616553 23.69407079 7.61405487 27.64654217 70.79034767 14.94629072
#> [29,] 24.55633696 24.14421560 7.74676017 28.18153790 72.23701660 15.22376416
#> [30,] 24.97792666 24.55658126 7.86815017 28.67179448 73.56404704 15.47784749
#>             [,7]        [,8]        [,9]      [,10]
#>  [1,] 0.02703816  0.04066922  0.02984621 0.02082147
#>  [2,] 0.10951554  0.17622920  0.12186052 0.08084208
#>  [3,] 0.26244968  0.44051781  0.29349771 0.18868116
#>  [4,] 0.48735607  0.84281868  0.54694387 0.34387423
#>  [5,] 0.77840909  1.37691907  0.87592642 0.54151555
#>  [6,] 1.12587705  2.02733525  1.26960202 0.77454490
#>  [7,] 1.51838666  2.77386040  1.71514989 1.03516538
#>  [8,] 1.94435879  3.59468368  2.19943162 1.31569698
#>  [9,] 2.39287008  4.46841007  2.70999723 1.60906069
#> [10,] 2.85412535  5.37527739  3.23564380 1.90902295
#> [11,] 3.31967215  6.29780891  3.76667591 2.21028719
#> [12,] 3.78244990  7.22108297  4.29497338 2.50849129
#> [13,] 4.23673782  8.13275318  4.81393996 2.80014998
#> [14,] 4.67804514  9.02291489  5.31838364 3.08256812
#> [15,] 5.10297284  9.88388482  5.80436230 3.35374161
#> [16,] 5.50906576 10.70993920  6.26901703 3.61225654
#> [17,] 5.89466714 11.49704058  6.71040695 3.85719277
#> [18,] 6.25878233 12.24257210  7.12735401 4.08803560
#> [19,] 6.60095557 12.94509043  7.51930213 4.30459712
#> [20,] 6.92116137 13.60410316  7.88619271 4.50694769
#> [21,] 7.21971057 14.21987317  8.22835686 4.69535738
#> [22,] 7.49717056 14.79324978  8.54642354 4.87024661
#> [23,] 7.75429844 15.32552558  8.84124267 5.03214521
#> [24,] 7.99198577 15.81831671  9.11382135 5.18165873
#> [25,] 8.21121355 16.27346404  9.36527185 5.31944128
#> [26,] 8.41301594 16.69295288  9.59676971 5.44617376
#> [27,] 8.59845151 17.07884844  9.80952037 5.56254675
#> [28,] 8.76858074 17.43324480 10.00473314 5.66924721
#> [29,] 8.92444883 17.75822529 10.18360117 5.76694850
#> [30,] 9.06707275 18.05583228 10.34728639 5.85630285