Access or assign maturity at age

This function can be used to access or assign the maturity at age in an lhm object.

Assignment can be of a list or numeric vector. When a list is assigned it should provide the \(a_{crit}\) and optionally the \(\delta\) parameter. If only acrit is supplied then maturity is knife-edged at that age. If both acrit and delta are supplied then maturity-at-age is calculated as a logistic ogive: $$Maturity = 1/(1+exp((acrit-Age)/delta)).$$

Uncertainty accross iterations in the lhm object can be included by providing a cv for each of the parameters. Parameter uncertainty is assumed to be log-normal with \(\sigma = \sqrt{ln(1+cv^2)}\).

When assigning a numeric vector it is assumed that each replicate is identical.

Usage

maturity(object, ...)

# S4 method for class 'lhm'
maturity(object)

maturity(object) <- value

# S4 method for class 'lhm,list'
maturity(object) <- value

# S4 method for class 'lhm,numeric'
maturity(object) <- value

Arguments

object

a lhm object

...

arguments to generic function

value

either a named list or numeric vector

numeric

a vector of length equal to the maximum age

list

a list object containing the components mu and cv

mu

a list containing the parameters acrit and (optionally) delta representing their mean values

cv

optional single numeric value or a list containing parameters acrit and delta representing their coefficients of variation

Value

Accessor function returns a matrix of maturity at age across iterations. Assignment function populates the matrix across iterations.

Examples

# initialize single iteration
dat <- lhm(ainf = 30, iter = 1)

# add knife-edged maturity
maturity(dat) <- list(mu = list(acrit = 10))
maturity(dat)
#>       [,1]
#>  [1,]    0
#>  [2,]    0
#>  [3,]    0
#>  [4,]    0
#>  [5,]    0
#>  [6,]    0
#>  [7,]    0
#>  [8,]    0
#>  [9,]    0
#> [10,]    1
#> [11,]    1
#> [12,]    1
#> [13,]    1
#> [14,]    1
#> [15,]    1
#> [16,]    1
#> [17,]    1
#> [18,]    1
#> [19,]    1
#> [20,]    1
#> [21,]    1
#> [22,]    1
#> [23,]    1
#> [24,]    1
#> [25,]    1
#> [26,]    1
#> [27,]    1
#> [28,]    1
#> [29,]    1
#> [30,]    1

# increase iterations
dat <- iterate(dat, iter = 10)
maturity(dat)
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#>  [1,]    0    0    0    0    0    0    0    0    0     0
#>  [2,]    0    0    0    0    0    0    0    0    0     0
#>  [3,]    0    0    0    0    0    0    0    0    0     0
#>  [4,]    0    0    0    0    0    0    0    0    0     0
#>  [5,]    0    0    0    0    0    0    0    0    0     0
#>  [6,]    0    0    0    0    0    0    0    0    0     0
#>  [7,]    0    0    0    0    0    0    0    0    0     0
#>  [8,]    0    0    0    0    0    0    0    0    0     0
#>  [9,]    0    0    0    0    0    0    0    0    0     0
#> [10,]    1    1    1    1    1    1    1    1    1     1
#> [11,]    1    1    1    1    1    1    1    1    1     1
#> [12,]    1    1    1    1    1    1    1    1    1     1
#> [13,]    1    1    1    1    1    1    1    1    1     1
#> [14,]    1    1    1    1    1    1    1    1    1     1
#> [15,]    1    1    1    1    1    1    1    1    1     1
#> [16,]    1    1    1    1    1    1    1    1    1     1
#> [17,]    1    1    1    1    1    1    1    1    1     1
#> [18,]    1    1    1    1    1    1    1    1    1     1
#> [19,]    1    1    1    1    1    1    1    1    1     1
#> [20,]    1    1    1    1    1    1    1    1    1     1
#> [21,]    1    1    1    1    1    1    1    1    1     1
#> [22,]    1    1    1    1    1    1    1    1    1     1
#> [23,]    1    1    1    1    1    1    1    1    1     1
#> [24,]    1    1    1    1    1    1    1    1    1     1
#> [25,]    1    1    1    1    1    1    1    1    1     1
#> [26,]    1    1    1    1    1    1    1    1    1     1
#> [27,]    1    1    1    1    1    1    1    1    1     1
#> [28,]    1    1    1    1    1    1    1    1    1     1
#> [29,]    1    1    1    1    1    1    1    1    1     1
#> [30,]    1    1    1    1    1    1    1    1    1     1

# add logistic maturity ogive
# with uncertainty
maturity(dat) <- list(mu=list(acrit = 1, delta = 3), 
                         cv=list(acrit = 0.01, delta = 0.01))
maturity(dat)
#>            [,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#>  [1,] 0.4992216 0.4998560 0.4998008 0.4986368 0.4999111 0.5001126 0.5015800
#>  [2,] 0.5817601 0.5829539 0.5824210 0.5814519 0.5821280 0.5804792 0.5840727
#>  [3,] 0.6599608 0.6615924 0.6606599 0.6599173 0.6600205 0.6567921 0.6621097
#>  [4,] 0.7303179 0.7322143 0.7310100 0.7304910 0.7301249 0.7257811 0.7322180
#>  [5,] 0.7907336 0.7927157 0.7913809 0.7910563 0.7903650 0.7854328 0.7923419
#>  [6,] 0.8405692 0.8424884 0.8411468 0.8409741 0.8401034 0.8350574 0.8418815
#>  [7,] 0.8803329 0.8820874 0.8808288 0.8807638 0.8798350 0.8750309 0.8813723
#>  [8,] 0.9112265 0.9127620 0.9116395 0.9116438 0.9107428 0.9064028 0.9120306
#>  [9,] 0.9347357 0.9360353 0.9350715 0.9351155 0.9342945 0.9305245 0.9353468
#> [10,] 0.9523448 0.9534167 0.9526126 0.9526757 0.9519596 0.9487808 0.9528028
#> [11,] 0.9653787 0.9662453 0.9655892 0.9656580 0.9650530 0.9624336 0.9657186
#> [12,] 0.9749416 0.9756313 0.9751052 0.9751717 0.9746728 0.9725524 0.9751919
#> [13,] 0.9819126 0.9824548 0.9820385 0.9820987 0.9816947 0.9800023 0.9820959
#> [14,] 0.9869702 0.9873923 0.9870665 0.9871186 0.9867962 0.9854603 0.9871039
#> [15,] 0.9906271 0.9909531 0.9907003 0.9907441 0.9904898 0.9894447 0.9907244
#> [16,] 0.9932647 0.9935148 0.9933201 0.9933561 0.9931573 0.9923457 0.9933353
#> [17,] 0.9951637 0.9953546 0.9952054 0.9952345 0.9950803 0.9944539 0.9952148
#> [18,] 0.9965291 0.9966742 0.9965605 0.9965837 0.9964648 0.9959838 0.9965662
#> [19,] 0.9975100 0.9976198 0.9975335 0.9975518 0.9974607 0.9970929 0.9975368
#> [20,] 0.9982142 0.9982970 0.9982318 0.9982461 0.9981765 0.9978963 0.9982336
#> [21,] 0.9987195 0.9987818 0.9987326 0.9987437 0.9986908 0.9984781 0.9987335
#> [22,] 0.9990819 0.9991287 0.9990917 0.9991003 0.9990602 0.9988991 0.9990921
#> [23,] 0.9993419 0.9993769 0.9993491 0.9993557 0.9993255 0.9992038 0.9993492
#> [24,] 0.9995282 0.9995544 0.9995336 0.9995387 0.9995159 0.9994242 0.9995335
#> [25,] 0.9996619 0.9996814 0.9996659 0.9996697 0.9996525 0.9995836 0.9996657
#> [26,] 0.9997576 0.9997722 0.9997606 0.9997635 0.9997507 0.9996989 0.9997604
#> [27,] 0.9998263 0.9998371 0.9998285 0.9998307 0.9998211 0.9997823 0.9998283
#> [28,] 0.9998755 0.9998835 0.9998771 0.9998788 0.9998716 0.9998426 0.9998769
#> [29,] 0.9999108 0.9999167 0.9999120 0.9999132 0.9999079 0.9998862 0.9999118
#> [30,] 0.9999360 0.9999404 0.9999369 0.9999379 0.9999339 0.9999177 0.9999368
#>            [,8]      [,9]     [,10]
#>  [1,] 0.5002352 0.5002646 0.4990928
#>  [2,] 0.5823360 0.5827359 0.5832563
#>  [3,] 0.6601136 0.6608246 0.6628317
#>  [4,] 0.7301197 0.7310443 0.7341386
#>  [5,] 0.7902876 0.7913172 0.7950285
#>  [6,] 0.8399814 0.8410204 0.8449187
#>  [7,] 0.8796923 0.8806706 0.8844308
#>  [8,] 0.9105973 0.9114728 0.9148899
#>  [9,] 0.9341575 0.9349118 0.9378851
#> [10,] 0.9518373 0.9524685 0.9549732
#> [11,] 0.9649479 0.9654645 0.9675231
#> [12,] 0.9745849 0.9750004 0.9766605
#> [13,] 0.9816228 0.9819525 0.9832716
#> [14,] 0.9867383 0.9869970 0.9880330
#> [15,] 0.9904438 0.9906450 0.9914509
#> [16,] 0.9931211 0.9932765 0.9938987
#> [17,] 0.9950520 0.9951714 0.9956487
#> [18,] 0.9964429 0.9965341 0.9968983
#> [19,] 0.9974438 0.9975132 0.9977898
#> [20,] 0.9981636 0.9982162 0.9984255
#> [21,] 0.9986810 0.9987207 0.9988786
#> [22,] 0.9990527 0.9990827 0.9992014
#> [23,] 0.9993198 0.9993423 0.9994313
#> [24,] 0.9995116 0.9995285 0.9995951
#> [25,] 0.9996493 0.9996620 0.9997117
#> [26,] 0.9997482 0.9997577 0.9997947
#> [27,] 0.9998192 0.9998263 0.9998539
#> [28,] 0.9998702 0.9998755 0.9998960
#> [29,] 0.9999068 0.9999107 0.9999259
#> [30,] 0.9999331 0.9999360 0.9999473