Access or assign natural mortality at age

This function can be used to access or assign the natural mortality at age in an lhm object.

Assignment can be of a list or numeric vector. When a list is assigned it should contain a value for mu, which is the mean natural mortality at age and may or may not be constant across ages.

If the list also contains cv then log-normal uncertainty is added around the mortality at age. If the list contains range then the uncertainty is bounded at these values. If range is provided but not cv, and if mortality is constant across ages, then a uniform uncertainy is implemented.

Usage

nmort(object, ...)

# S4 method for class 'lhm'
nmort(object)

nmort(object) <- value

# S4 method for class 'lhm,list'
nmort(object) <- value

# S4 method for class 'lhm,numeric'
nmort(object) <- value

Arguments

object

a lhm object

...

arguments to generic function

value

either a named list or numeric vector

numeric

a single numeric or vector of values

list

a list object containing the components mu, cv and range

mu

a single numeric or vector of values representing the mean natural mortality

cv

optional single numeric value representing the coefficient of variation

range

optional vector of length two containing the lower and upper bounds

Value

Accessor function returns a matrix of maturity at age across iterations. Assignment function populates the matrix across iterations.

Examples

# M at age vector input
dat <- lhm(ainf = 10,iter=1)
nmort(dat) <- c(0.1,0.1,0.2)
nmort(dat)
#>       [,1]
#>  [1,]  0.1
#>  [2,]  0.1
#>  [3,]  0.2
#>  [4,]  0.2
#>  [5,]  0.2
#>  [6,]  0.2
#>  [7,]  0.2
#>  [8,]  0.2
#>  [9,]  0.2
#> [10,]  0.2

# constant M with log-normal uncertainty
dat <- lhm(ainf=10, iter = 20)
nmort(dat) <- list(mu = 0.2, cv = 0.25)
nmort(dat)
#>            [,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#>  [1,] 0.2399111 0.1827493 0.1844286 0.1949469 0.1954459 0.2221569 0.1108381
#>  [2,] 0.2399111 0.1827493 0.1844286 0.1949469 0.1954459 0.2221569 0.1108381
#>  [3,] 0.2399111 0.1827493 0.1844286 0.1949469 0.1954459 0.2221569 0.1108381
#>  [4,] 0.2399111 0.1827493 0.1844286 0.1949469 0.1954459 0.2221569 0.1108381
#>  [5,] 0.2399111 0.1827493 0.1844286 0.1949469 0.1954459 0.2221569 0.1108381
#>  [6,] 0.2399111 0.1827493 0.1844286 0.1949469 0.1954459 0.2221569 0.1108381
#>  [7,] 0.2399111 0.1827493 0.1844286 0.1949469 0.1954459 0.2221569 0.1108381
#>  [8,] 0.2399111 0.1827493 0.1844286 0.1949469 0.1954459 0.2221569 0.1108381
#>  [9,] 0.2399111 0.1827493 0.1844286 0.1949469 0.1954459 0.2221569 0.1108381
#> [10,] 0.2399111 0.1827493 0.1844286 0.1949469 0.1954459 0.2221569 0.1108381
#>           [,8]      [,9]     [,10]    [,11]     [,12]     [,13]     [,14]
#>  [1,] 0.375593 0.1775167 0.2044938 0.252783 0.1647199 0.2552601 0.1826296
#>  [2,] 0.375593 0.1775167 0.2044938 0.252783 0.1647199 0.2552601 0.1826296
#>  [3,] 0.375593 0.1775167 0.2044938 0.252783 0.1647199 0.2552601 0.1826296
#>  [4,] 0.375593 0.1775167 0.2044938 0.252783 0.1647199 0.2552601 0.1826296
#>  [5,] 0.375593 0.1775167 0.2044938 0.252783 0.1647199 0.2552601 0.1826296
#>  [6,] 0.375593 0.1775167 0.2044938 0.252783 0.1647199 0.2552601 0.1826296
#>  [7,] 0.375593 0.1775167 0.2044938 0.252783 0.1647199 0.2552601 0.1826296
#>  [8,] 0.375593 0.1775167 0.2044938 0.252783 0.1647199 0.2552601 0.1826296
#>  [9,] 0.375593 0.1775167 0.2044938 0.252783 0.1647199 0.2552601 0.1826296
#> [10,] 0.375593 0.1775167 0.2044938 0.252783 0.1647199 0.2552601 0.1826296
#>           [,15]     [,16]     [,17]    [,18]     [,19]     [,20]
#>  [1,] 0.1451931 0.1525863 0.2522055 0.200422 0.2188345 0.1276849
#>  [2,] 0.1451931 0.1525863 0.2522055 0.200422 0.2188345 0.1276849
#>  [3,] 0.1451931 0.1525863 0.2522055 0.200422 0.2188345 0.1276849
#>  [4,] 0.1451931 0.1525863 0.2522055 0.200422 0.2188345 0.1276849
#>  [5,] 0.1451931 0.1525863 0.2522055 0.200422 0.2188345 0.1276849
#>  [6,] 0.1451931 0.1525863 0.2522055 0.200422 0.2188345 0.1276849
#>  [7,] 0.1451931 0.1525863 0.2522055 0.200422 0.2188345 0.1276849
#>  [8,] 0.1451931 0.1525863 0.2522055 0.200422 0.2188345 0.1276849
#>  [9,] 0.1451931 0.1525863 0.2522055 0.200422 0.2188345 0.1276849
#> [10,] 0.1451931 0.1525863 0.2522055 0.200422 0.2188345 0.1276849

# constant M with bounded log-normal uncertainty
dat <- lhm(ainf = 10, iter = 20)
nmort(dat) <- list(mu = 0.2, cv = 0.5, range = c(0.1,0.3))
nmort(dat)
#>       [,1]      [,2] [,3]      [,4] [,5]      [,6]      [,7]      [,8]
#>  [1,]  0.1 0.2045828  0.3 0.2000478  0.3 0.2291165 0.2382598 0.1698327
#>  [2,]  0.1 0.2045828  0.3 0.2000478  0.3 0.2291165 0.2382598 0.1698327
#>  [3,]  0.1 0.2045828  0.3 0.2000478  0.3 0.2291165 0.2382598 0.1698327
#>  [4,]  0.1 0.2045828  0.3 0.2000478  0.3 0.2291165 0.2382598 0.1698327
#>  [5,]  0.1 0.2045828  0.3 0.2000478  0.3 0.2291165 0.2382598 0.1698327
#>  [6,]  0.1 0.2045828  0.3 0.2000478  0.3 0.2291165 0.2382598 0.1698327
#>  [7,]  0.1 0.2045828  0.3 0.2000478  0.3 0.2291165 0.2382598 0.1698327
#>  [8,]  0.1 0.2045828  0.3 0.2000478  0.3 0.2291165 0.2382598 0.1698327
#>  [9,]  0.1 0.2045828  0.3 0.2000478  0.3 0.2291165 0.2382598 0.1698327
#> [10,]  0.1 0.2045828  0.3 0.2000478  0.3 0.2291165 0.2382598 0.1698327
#>            [,9]     [,10] [,11] [,12]     [,13]     [,14]     [,15] [,16] [,17]
#>  [1,] 0.1940432 0.1714131   0.3   0.3 0.2547952 0.2326407 0.1380698   0.3   0.1
#>  [2,] 0.1940432 0.1714131   0.3   0.3 0.2547952 0.2326407 0.1380698   0.3   0.1
#>  [3,] 0.1940432 0.1714131   0.3   0.3 0.2547952 0.2326407 0.1380698   0.3   0.1
#>  [4,] 0.1940432 0.1714131   0.3   0.3 0.2547952 0.2326407 0.1380698   0.3   0.1
#>  [5,] 0.1940432 0.1714131   0.3   0.3 0.2547952 0.2326407 0.1380698   0.3   0.1
#>  [6,] 0.1940432 0.1714131   0.3   0.3 0.2547952 0.2326407 0.1380698   0.3   0.1
#>  [7,] 0.1940432 0.1714131   0.3   0.3 0.2547952 0.2326407 0.1380698   0.3   0.1
#>  [8,] 0.1940432 0.1714131   0.3   0.3 0.2547952 0.2326407 0.1380698   0.3   0.1
#>  [9,] 0.1940432 0.1714131   0.3   0.3 0.2547952 0.2326407 0.1380698   0.3   0.1
#> [10,] 0.1940432 0.1714131   0.3   0.3 0.2547952 0.2326407 0.1380698   0.3   0.1
#>           [,18]     [,19]    [,20]
#>  [1,] 0.1662012 0.2195778 0.149354
#>  [2,] 0.1662012 0.2195778 0.149354
#>  [3,] 0.1662012 0.2195778 0.149354
#>  [4,] 0.1662012 0.2195778 0.149354
#>  [5,] 0.1662012 0.2195778 0.149354
#>  [6,] 0.1662012 0.2195778 0.149354
#>  [7,] 0.1662012 0.2195778 0.149354
#>  [8,] 0.1662012 0.2195778 0.149354
#>  [9,] 0.1662012 0.2195778 0.149354
#> [10,] 0.1662012 0.2195778 0.149354

# constant M with uniform uncertainty
dat <- lhm(ainf = 10,iter = 20)
nmort(dat) <- list(range = c(0.1,0.3))
nmort(dat)
#>            [,1]      [,2]      [,3]      [,4]     [,5]      [,6]      [,7]
#>  [1,] 0.2328571 0.1170449 0.2712264 0.1153967 0.270569 0.1212694 0.1969606
#>  [2,] 0.2328571 0.1170449 0.2712264 0.1153967 0.270569 0.1212694 0.1969606
#>  [3,] 0.2328571 0.1170449 0.2712264 0.1153967 0.270569 0.1212694 0.1969606
#>  [4,] 0.2328571 0.1170449 0.2712264 0.1153967 0.270569 0.1212694 0.1969606
#>  [5,] 0.2328571 0.1170449 0.2712264 0.1153967 0.270569 0.1212694 0.1969606
#>  [6,] 0.2328571 0.1170449 0.2712264 0.1153967 0.270569 0.1212694 0.1969606
#>  [7,] 0.2328571 0.1170449 0.2712264 0.1153967 0.270569 0.1212694 0.1969606
#>  [8,] 0.2328571 0.1170449 0.2712264 0.1153967 0.270569 0.1212694 0.1969606
#>  [9,] 0.2328571 0.1170449 0.2712264 0.1153967 0.270569 0.1212694 0.1969606
#> [10,] 0.2328571 0.1170449 0.2712264 0.1153967 0.270569 0.1212694 0.1969606
#>            [,8]      [,9]     [,10]    [,11]     [,12]     [,13]     [,14]
#>  [1,] 0.1494438 0.2373138 0.1327246 0.290565 0.1643709 0.1723068 0.2775447
#>  [2,] 0.1494438 0.2373138 0.1327246 0.290565 0.1643709 0.1723068 0.2775447
#>  [3,] 0.1494438 0.2373138 0.1327246 0.290565 0.1643709 0.1723068 0.2775447
#>  [4,] 0.1494438 0.2373138 0.1327246 0.290565 0.1643709 0.1723068 0.2775447
#>  [5,] 0.1494438 0.2373138 0.1327246 0.290565 0.1643709 0.1723068 0.2775447
#>  [6,] 0.1494438 0.2373138 0.1327246 0.290565 0.1643709 0.1723068 0.2775447
#>  [7,] 0.1494438 0.2373138 0.1327246 0.290565 0.1643709 0.1723068 0.2775447
#>  [8,] 0.1494438 0.2373138 0.1327246 0.290565 0.1643709 0.1723068 0.2775447
#>  [9,] 0.1494438 0.2373138 0.1327246 0.290565 0.1643709 0.1723068 0.2775447
#> [10,] 0.1494438 0.2373138 0.1327246 0.290565 0.1643709 0.1723068 0.2775447
#>           [,15]     [,16]     [,17]     [,18]     [,19]     [,20]
#>  [1,] 0.2656029 0.1201313 0.2812103 0.2545461 0.1766741 0.2999305
#>  [2,] 0.2656029 0.1201313 0.2812103 0.2545461 0.1766741 0.2999305
#>  [3,] 0.2656029 0.1201313 0.2812103 0.2545461 0.1766741 0.2999305
#>  [4,] 0.2656029 0.1201313 0.2812103 0.2545461 0.1766741 0.2999305
#>  [5,] 0.2656029 0.1201313 0.2812103 0.2545461 0.1766741 0.2999305
#>  [6,] 0.2656029 0.1201313 0.2812103 0.2545461 0.1766741 0.2999305
#>  [7,] 0.2656029 0.1201313 0.2812103 0.2545461 0.1766741 0.2999305
#>  [8,] 0.2656029 0.1201313 0.2812103 0.2545461 0.1766741 0.2999305
#>  [9,] 0.2656029 0.1201313 0.2812103 0.2545461 0.1766741 0.2999305
#> [10,] 0.2656029 0.1201313 0.2812103 0.2545461 0.1766741 0.2999305

# M at age with log-normal uncertainty
dat <- lhm(ainf=10, iter = 20)
nmort(dat) <- list(mu = c(0.1,0.1,0.2), cv = 0.25)
nmort(dat)
#>             [,1]       [,2]       [,3]       [,4]      [,5]       [,6]
#>  [1,] 0.08819214 0.07995533 0.07478875 0.07975657 0.0629554 0.08184549
#>  [2,] 0.08819214 0.07995533 0.07478875 0.07975657 0.0629554 0.08184549
#>  [3,] 0.17638429 0.15991066 0.14957749 0.15951314 0.1259108 0.16369097
#>  [4,] 0.17638429 0.15991066 0.14957749 0.15951314 0.1259108 0.16369097
#>  [5,] 0.17638429 0.15991066 0.14957749 0.15951314 0.1259108 0.16369097
#>  [6,] 0.17638429 0.15991066 0.14957749 0.15951314 0.1259108 0.16369097
#>  [7,] 0.17638429 0.15991066 0.14957749 0.15951314 0.1259108 0.16369097
#>  [8,] 0.17638429 0.15991066 0.14957749 0.15951314 0.1259108 0.16369097
#>  [9,] 0.17638429 0.15991066 0.14957749 0.15951314 0.1259108 0.16369097
#> [10,] 0.17638429 0.15991066 0.14957749 0.15951314 0.1259108 0.16369097
#>             [,7]       [,8]      [,9]     [,10]      [,11]      [,12]     [,13]
#>  [1,] 0.08454917 0.08500587 0.1025942 0.1234423 0.09215083 0.06873744 0.1033907
#>  [2,] 0.08454917 0.08500587 0.1025942 0.1234423 0.09215083 0.06873744 0.1033907
#>  [3,] 0.16909834 0.17001173 0.2051884 0.2468846 0.18430165 0.13747489 0.2067814
#>  [4,] 0.16909834 0.17001173 0.2051884 0.2468846 0.18430165 0.13747489 0.2067814
#>  [5,] 0.16909834 0.17001173 0.2051884 0.2468846 0.18430165 0.13747489 0.2067814
#>  [6,] 0.16909834 0.17001173 0.2051884 0.2468846 0.18430165 0.13747489 0.2067814
#>  [7,] 0.16909834 0.17001173 0.2051884 0.2468846 0.18430165 0.13747489 0.2067814
#>  [8,] 0.16909834 0.17001173 0.2051884 0.2468846 0.18430165 0.13747489 0.2067814
#>  [9,] 0.16909834 0.17001173 0.2051884 0.2468846 0.18430165 0.13747489 0.2067814
#> [10,] 0.16909834 0.17001173 0.2051884 0.2468846 0.18430165 0.13747489 0.2067814
#>            [,14]     [,15]     [,16]     [,17]      [,18]     [,19]     [,20]
#>  [1,] 0.08701476 0.1115931 0.1637814 0.1077126 0.06408132 0.1031584 0.1263446
#>  [2,] 0.08701476 0.1115931 0.1637814 0.1077126 0.06408132 0.1031584 0.1263446
#>  [3,] 0.17402951 0.2231861 0.3275627 0.2154253 0.12816264 0.2063168 0.2526891
#>  [4,] 0.17402951 0.2231861 0.3275627 0.2154253 0.12816264 0.2063168 0.2526891
#>  [5,] 0.17402951 0.2231861 0.3275627 0.2154253 0.12816264 0.2063168 0.2526891
#>  [6,] 0.17402951 0.2231861 0.3275627 0.2154253 0.12816264 0.2063168 0.2526891
#>  [7,] 0.17402951 0.2231861 0.3275627 0.2154253 0.12816264 0.2063168 0.2526891
#>  [8,] 0.17402951 0.2231861 0.3275627 0.2154253 0.12816264 0.2063168 0.2526891
#>  [9,] 0.17402951 0.2231861 0.3275627 0.2154253 0.12816264 0.2063168 0.2526891
#> [10,] 0.17402951 0.2231861 0.3275627 0.2154253 0.12816264 0.2063168 0.2526891

# M at age with bounded log-normal uncertainty
dat <- lhm(ainf = 10, iter = 20)
nmort(dat) <- list(mu = c(0.1,0.1,0.2), cv = 0.25, range = c(0.1,0.3))
nmort(dat)
#>            [,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#>  [1,] 0.1602913 0.1083669 0.1366675 0.1077579 0.1000000 0.1000000 0.1125592
#>  [2,] 0.1602913 0.1083669 0.1366675 0.1077579 0.1000000 0.1000000 0.1125592
#>  [3,] 0.3000000 0.2167338 0.2733351 0.2155159 0.1992369 0.1950965 0.2251184
#>  [4,] 0.3000000 0.2167338 0.2733351 0.2155159 0.1992369 0.1950965 0.2251184
#>  [5,] 0.3000000 0.2167338 0.2733351 0.2155159 0.1992369 0.1950965 0.2251184
#>  [6,] 0.3000000 0.2167338 0.2733351 0.2155159 0.1992369 0.1950965 0.2251184
#>  [7,] 0.3000000 0.2167338 0.2733351 0.2155159 0.1992369 0.1950965 0.2251184
#>  [8,] 0.3000000 0.2167338 0.2733351 0.2155159 0.1992369 0.1950965 0.2251184
#>  [9,] 0.3000000 0.2167338 0.2733351 0.2155159 0.1992369 0.1950965 0.2251184
#> [10,] 0.3000000 0.2167338 0.2733351 0.2155159 0.1992369 0.1950965 0.2251184
#>            [,8]      [,9]     [,10]     [,11]     [,12]     [,13]    [,14]
#>  [1,] 0.1000000 0.1000000 0.1017144 0.1000000 0.1194868 0.1000000 0.100000
#>  [2,] 0.1000000 0.1000000 0.1017144 0.1000000 0.1194868 0.1000000 0.100000
#>  [3,] 0.1818778 0.1703637 0.2034288 0.1463187 0.2389736 0.1979721 0.140704
#>  [4,] 0.1818778 0.1703637 0.2034288 0.1463187 0.2389736 0.1979721 0.140704
#>  [5,] 0.1818778 0.1703637 0.2034288 0.1463187 0.2389736 0.1979721 0.140704
#>  [6,] 0.1818778 0.1703637 0.2034288 0.1463187 0.2389736 0.1979721 0.140704
#>  [7,] 0.1818778 0.1703637 0.2034288 0.1463187 0.2389736 0.1979721 0.140704
#>  [8,] 0.1818778 0.1703637 0.2034288 0.1463187 0.2389736 0.1979721 0.140704
#>  [9,] 0.1818778 0.1703637 0.2034288 0.1463187 0.2389736 0.1979721 0.140704
#> [10,] 0.1818778 0.1703637 0.2034288 0.1463187 0.2389736 0.1979721 0.140704
#>           [,15]     [,16]     [,17]     [,18]     [,19]     [,20]
#>  [1,] 0.1000000 0.1084974 0.1000000 0.1000000 0.1209305 0.1000000
#>  [2,] 0.1000000 0.1084974 0.1000000 0.1000000 0.1209305 0.1000000
#>  [3,] 0.1537531 0.2169948 0.1571871 0.1807953 0.2418610 0.1972707
#>  [4,] 0.1537531 0.2169948 0.1571871 0.1807953 0.2418610 0.1972707
#>  [5,] 0.1537531 0.2169948 0.1571871 0.1807953 0.2418610 0.1972707
#>  [6,] 0.1537531 0.2169948 0.1571871 0.1807953 0.2418610 0.1972707
#>  [7,] 0.1537531 0.2169948 0.1571871 0.1807953 0.2418610 0.1972707
#>  [8,] 0.1537531 0.2169948 0.1571871 0.1807953 0.2418610 0.1972707
#>  [9,] 0.1537531 0.2169948 0.1571871 0.1807953 0.2418610 0.1972707
#> [10,] 0.1537531 0.2169948 0.1571871 0.1807953 0.2418610 0.1972707