Access or assign mass at age

This function can be used to access or assign the biomass at age in an lhm object.

Assignment can be of a list or numeric vector. When a list is assigned it should provide the \(a\) and \(b\) parameters of the function: $$Mass = a * Length ^ b.$$

Uncertainty accross iterations in the lhm object can be included by providing either an overall coefficient of variation (cv) or a cv for each of the parameters. Overall uncertainty is assumed to be log-normal with \(\sigma = \sqrt{ln(1+cv^2)}\). Parameter uncertainty is assumed to be normal with \(\sigma = \mu * cv\).

When assigning a numeric vector it is assumed that each replicate is identical.

Usage

mass(object, ...)

# S4 method for class 'lhm'
mass(object)

mass(object) <- value

# S4 method for class 'lhm,list'
mass(object) <- value

# S4 method for class 'lhm,numeric'
mass(object) <- value

Arguments

object

a lhm object

...

arguments to generic function

value

either a named list or numeric vector

numeric

a vector of length equal to the maximum age

list

a list object containing the components mu and cv

mu

a list containing the parameters a and b representing their mean values

cv

optional single numeric value or a list containing parameters a and b representing their coefficients of variation

Value

Accessor function returns a matrix of mass at age across iterations. Assignment function populates the matrix across iterations.

Examples

# initialize single iteration
dat <- lhm(ainf = 30, iter = 1)

# add length data
size(dat) <- list(mu=list(Linf = 100, k = 0.1, t0 = -0.5))
size(dat)
#>           [,1]
#>  [1,] 13.92920
#>  [2,] 22.11992
#>  [3,] 29.53119
#>  [4,] 36.23718
#>  [5,] 42.30502
#>  [6,] 47.79542
#>  [7,] 52.76334
#>  [8,] 57.25851
#>  [9,] 61.32590
#> [10,] 65.00623
#> [11,] 68.33632
#> [12,] 71.34952
#> [13,] 74.07597
#> [14,] 76.54297
#> [15,] 78.77520
#> [16,] 80.79501
#> [17,] 82.62261
#> [18,] 84.27628
#> [19,] 85.77259
#> [20,] 87.12651
#> [21,] 88.35158
#> [22,] 89.46008
#> [23,] 90.46308
#> [24,] 91.37064
#> [25,] 92.19183
#> [26,] 92.93488
#> [27,] 93.60721
#> [28,] 94.21557
#> [29,] 94.76603
#> [30,] 95.26411

# mass at age
mass(dat) <- list(mu = list(a = 9.63e-6, b = 3.173))
mass(dat)
#>              [,1]
#>  [1,]  0.04104916
#>  [2,]  0.17808388
#>  [3,]  0.44547956
#>  [4,]  0.85275321
#>  [5,]  1.39369508
#>  [6,]  2.05266949
#>  [7,]  2.80922635
#>  [8,]  3.64126817
#>  [9,]  4.52710352
#> [10,]  5.44668595
#> [11,]  6.38227976
#> [12,]  7.31873653
#> [13,]  8.24351785
#> [14,]  9.14656144
#> [15,] 10.02005862
#> [16,] 10.85818945
#> [17,] 11.65684592
#> [18,] 12.41336259
#> [19,] 13.12626598
#> [20,] 13.79504878
#> [21,] 14.41997116
#> [22,] 15.00188948
#> [23,] 15.54211085
#> [24,] 16.04227148
#> [25,] 16.50423644
#> [26,] 16.93001792
#> [27,] 17.32170973
#> [28,] 17.68143546
#> [29,] 18.01130809
#> [30,] 18.31339938

# iterate
dat <- iterate(dat, iter = 10)
size(dat)
#>           [,1]     [,2]     [,3]     [,4]     [,5]     [,6]     [,7]     [,8]
#>  [1,] 13.92920 13.92920 13.92920 13.92920 13.92920 13.92920 13.92920 13.92920
#>  [2,] 22.11992 22.11992 22.11992 22.11992 22.11992 22.11992 22.11992 22.11992
#>  [3,] 29.53119 29.53119 29.53119 29.53119 29.53119 29.53119 29.53119 29.53119
#>  [4,] 36.23718 36.23718 36.23718 36.23718 36.23718 36.23718 36.23718 36.23718
#>  [5,] 42.30502 42.30502 42.30502 42.30502 42.30502 42.30502 42.30502 42.30502
#>  [6,] 47.79542 47.79542 47.79542 47.79542 47.79542 47.79542 47.79542 47.79542
#>  [7,] 52.76334 52.76334 52.76334 52.76334 52.76334 52.76334 52.76334 52.76334
#>  [8,] 57.25851 57.25851 57.25851 57.25851 57.25851 57.25851 57.25851 57.25851
#>  [9,] 61.32590 61.32590 61.32590 61.32590 61.32590 61.32590 61.32590 61.32590
#> [10,] 65.00623 65.00623 65.00623 65.00623 65.00623 65.00623 65.00623 65.00623
#> [11,] 68.33632 68.33632 68.33632 68.33632 68.33632 68.33632 68.33632 68.33632
#> [12,] 71.34952 71.34952 71.34952 71.34952 71.34952 71.34952 71.34952 71.34952
#> [13,] 74.07597 74.07597 74.07597 74.07597 74.07597 74.07597 74.07597 74.07597
#> [14,] 76.54297 76.54297 76.54297 76.54297 76.54297 76.54297 76.54297 76.54297
#> [15,] 78.77520 78.77520 78.77520 78.77520 78.77520 78.77520 78.77520 78.77520
#> [16,] 80.79501 80.79501 80.79501 80.79501 80.79501 80.79501 80.79501 80.79501
#> [17,] 82.62261 82.62261 82.62261 82.62261 82.62261 82.62261 82.62261 82.62261
#> [18,] 84.27628 84.27628 84.27628 84.27628 84.27628 84.27628 84.27628 84.27628
#> [19,] 85.77259 85.77259 85.77259 85.77259 85.77259 85.77259 85.77259 85.77259
#> [20,] 87.12651 87.12651 87.12651 87.12651 87.12651 87.12651 87.12651 87.12651
#> [21,] 88.35158 88.35158 88.35158 88.35158 88.35158 88.35158 88.35158 88.35158
#> [22,] 89.46008 89.46008 89.46008 89.46008 89.46008 89.46008 89.46008 89.46008
#> [23,] 90.46308 90.46308 90.46308 90.46308 90.46308 90.46308 90.46308 90.46308
#> [24,] 91.37064 91.37064 91.37064 91.37064 91.37064 91.37064 91.37064 91.37064
#> [25,] 92.19183 92.19183 92.19183 92.19183 92.19183 92.19183 92.19183 92.19183
#> [26,] 92.93488 92.93488 92.93488 92.93488 92.93488 92.93488 92.93488 92.93488
#> [27,] 93.60721 93.60721 93.60721 93.60721 93.60721 93.60721 93.60721 93.60721
#> [28,] 94.21557 94.21557 94.21557 94.21557 94.21557 94.21557 94.21557 94.21557
#> [29,] 94.76603 94.76603 94.76603 94.76603 94.76603 94.76603 94.76603 94.76603
#> [30,] 95.26411 95.26411 95.26411 95.26411 95.26411 95.26411 95.26411 95.26411
#>           [,9]    [,10]
#>  [1,] 13.92920 13.92920
#>  [2,] 22.11992 22.11992
#>  [3,] 29.53119 29.53119
#>  [4,] 36.23718 36.23718
#>  [5,] 42.30502 42.30502
#>  [6,] 47.79542 47.79542
#>  [7,] 52.76334 52.76334
#>  [8,] 57.25851 57.25851
#>  [9,] 61.32590 61.32590
#> [10,] 65.00623 65.00623
#> [11,] 68.33632 68.33632
#> [12,] 71.34952 71.34952
#> [13,] 74.07597 74.07597
#> [14,] 76.54297 76.54297
#> [15,] 78.77520 78.77520
#> [16,] 80.79501 80.79501
#> [17,] 82.62261 82.62261
#> [18,] 84.27628 84.27628
#> [19,] 85.77259 85.77259
#> [20,] 87.12651 87.12651
#> [21,] 88.35158 88.35158
#> [22,] 89.46008 89.46008
#> [23,] 90.46308 90.46308
#> [24,] 91.37064 91.37064
#> [25,] 92.19183 92.19183
#> [26,] 92.93488 92.93488
#> [27,] 93.60721 93.60721
#> [28,] 94.21557 94.21557
#> [29,] 94.76603 94.76603
#> [30,] 95.26411 95.26411
mass(dat)
#>              [,1]        [,2]        [,3]        [,4]        [,5]        [,6]
#>  [1,]  0.04104916  0.04104916  0.04104916  0.04104916  0.04104916  0.04104916
#>  [2,]  0.17808388  0.17808388  0.17808388  0.17808388  0.17808388  0.17808388
#>  [3,]  0.44547956  0.44547956  0.44547956  0.44547956  0.44547956  0.44547956
#>  [4,]  0.85275321  0.85275321  0.85275321  0.85275321  0.85275321  0.85275321
#>  [5,]  1.39369508  1.39369508  1.39369508  1.39369508  1.39369508  1.39369508
#>  [6,]  2.05266949  2.05266949  2.05266949  2.05266949  2.05266949  2.05266949
#>  [7,]  2.80922635  2.80922635  2.80922635  2.80922635  2.80922635  2.80922635
#>  [8,]  3.64126817  3.64126817  3.64126817  3.64126817  3.64126817  3.64126817
#>  [9,]  4.52710352  4.52710352  4.52710352  4.52710352  4.52710352  4.52710352
#> [10,]  5.44668595  5.44668595  5.44668595  5.44668595  5.44668595  5.44668595
#> [11,]  6.38227976  6.38227976  6.38227976  6.38227976  6.38227976  6.38227976
#> [12,]  7.31873653  7.31873653  7.31873653  7.31873653  7.31873653  7.31873653
#> [13,]  8.24351785  8.24351785  8.24351785  8.24351785  8.24351785  8.24351785
#> [14,]  9.14656144  9.14656144  9.14656144  9.14656144  9.14656144  9.14656144
#> [15,] 10.02005862 10.02005862 10.02005862 10.02005862 10.02005862 10.02005862
#> [16,] 10.85818945 10.85818945 10.85818945 10.85818945 10.85818945 10.85818945
#> [17,] 11.65684592 11.65684592 11.65684592 11.65684592 11.65684592 11.65684592
#> [18,] 12.41336259 12.41336259 12.41336259 12.41336259 12.41336259 12.41336259
#> [19,] 13.12626598 13.12626598 13.12626598 13.12626598 13.12626598 13.12626598
#> [20,] 13.79504878 13.79504878 13.79504878 13.79504878 13.79504878 13.79504878
#> [21,] 14.41997116 14.41997116 14.41997116 14.41997116 14.41997116 14.41997116
#> [22,] 15.00188948 15.00188948 15.00188948 15.00188948 15.00188948 15.00188948
#> [23,] 15.54211085 15.54211085 15.54211085 15.54211085 15.54211085 15.54211085
#> [24,] 16.04227148 16.04227148 16.04227148 16.04227148 16.04227148 16.04227148
#> [25,] 16.50423644 16.50423644 16.50423644 16.50423644 16.50423644 16.50423644
#> [26,] 16.93001792 16.93001792 16.93001792 16.93001792 16.93001792 16.93001792
#> [27,] 17.32170973 17.32170973 17.32170973 17.32170973 17.32170973 17.32170973
#> [28,] 17.68143546 17.68143546 17.68143546 17.68143546 17.68143546 17.68143546
#> [29,] 18.01130809 18.01130809 18.01130809 18.01130809 18.01130809 18.01130809
#> [30,] 18.31339938 18.31339938 18.31339938 18.31339938 18.31339938 18.31339938
#>              [,7]        [,8]        [,9]       [,10]
#>  [1,]  0.04104916  0.04104916  0.04104916  0.04104916
#>  [2,]  0.17808388  0.17808388  0.17808388  0.17808388
#>  [3,]  0.44547956  0.44547956  0.44547956  0.44547956
#>  [4,]  0.85275321  0.85275321  0.85275321  0.85275321
#>  [5,]  1.39369508  1.39369508  1.39369508  1.39369508
#>  [6,]  2.05266949  2.05266949  2.05266949  2.05266949
#>  [7,]  2.80922635  2.80922635  2.80922635  2.80922635
#>  [8,]  3.64126817  3.64126817  3.64126817  3.64126817
#>  [9,]  4.52710352  4.52710352  4.52710352  4.52710352
#> [10,]  5.44668595  5.44668595  5.44668595  5.44668595
#> [11,]  6.38227976  6.38227976  6.38227976  6.38227976
#> [12,]  7.31873653  7.31873653  7.31873653  7.31873653
#> [13,]  8.24351785  8.24351785  8.24351785  8.24351785
#> [14,]  9.14656144  9.14656144  9.14656144  9.14656144
#> [15,] 10.02005862 10.02005862 10.02005862 10.02005862
#> [16,] 10.85818945 10.85818945 10.85818945 10.85818945
#> [17,] 11.65684592 11.65684592 11.65684592 11.65684592
#> [18,] 12.41336259 12.41336259 12.41336259 12.41336259
#> [19,] 13.12626598 13.12626598 13.12626598 13.12626598
#> [20,] 13.79504878 13.79504878 13.79504878 13.79504878
#> [21,] 14.41997116 14.41997116 14.41997116 14.41997116
#> [22,] 15.00188948 15.00188948 15.00188948 15.00188948
#> [23,] 15.54211085 15.54211085 15.54211085 15.54211085
#> [24,] 16.04227148 16.04227148 16.04227148 16.04227148
#> [25,] 16.50423644 16.50423644 16.50423644 16.50423644
#> [26,] 16.93001792 16.93001792 16.93001792 16.93001792
#> [27,] 17.32170973 17.32170973 17.32170973 17.32170973
#> [28,] 17.68143546 17.68143546 17.68143546 17.68143546
#> [29,] 18.01130809 18.01130809 18.01130809 18.01130809
#> [30,] 18.31339938 18.31339938 18.31339938 18.31339938

# multiple stochastic iterations
# with overall uncertainty
mass(dat) <- list(mu = list(a = 9.63e-6,b = 3.173), cv = 0.05)
mass(dat)
#>              [,1]       [,2]        [,3]        [,4]        [,5]        [,6]
#>  [1,]  0.04001061  0.0459552  0.05162686  0.02678824  0.03793009  0.03795629
#>  [2,]  0.17357833  0.1993678  0.22397321  0.11621564  0.16455241  0.16466606
#>  [3,]  0.43420886  0.4987216  0.56027242  0.29071522  0.41163039  0.41191469
#>  [4,]  0.83117843  0.9546710  1.07249388  0.55649767  0.78795790  0.78850211
#>  [5,]  1.35843439  1.5602641  1.75282771  0.90951058  1.28779703  1.28868646
#>  [6,]  2.00073666  2.2979966  2.58160914  1.33955027  1.89670016  1.89801013
#>  [7,]  2.73815251  3.1449742  3.53311843  1.83327123  2.59577105  2.59756384
#>  [8,]  3.54914356  4.0764584  4.57956395  2.37625287  3.36459130  3.36691508
#>  [9,]  4.41256715  5.0681654  5.69366471  2.95433960  4.18311764  4.18600674
#> [10,]  5.30888401  6.0976528  6.85020862  3.55444930  5.03282683  5.03630279
#> [11,]  6.22080716  7.1450651  8.02688978  4.16500787  5.89733080  5.90140383
#> [12,]  7.13357144  8.1934436  9.20465628  4.77612960  6.76263216  6.76730282
#> [13,]  8.03495566  9.2287512 10.36773876  5.37963205  7.61714522  7.62240605
#> [14,]  8.91515212 10.2397231 11.50348204  5.96894868  8.45157225  8.45740939
#> [15,]  9.76654969 11.2176172 12.60206528  6.53898365  9.25869792  9.26509251
#> [16,] 10.58347569 12.1559182 13.65616884  7.08593891 10.03314450 10.04007397
#> [17,] 11.36192604 13.0500270 14.66062613  7.60713363 10.77111613 10.77855528
#> [18,] 12.09930273 13.8969596 15.61208489  8.10082836 11.47014990 11.47807184
#> [19,] 12.79416956 14.6950665 16.50869193  8.56606153 12.12888429 12.13726119
#> [20,] 13.44603205 15.4437796 17.34980920  9.00250207 12.74685052 12.75565422
#> [21,] 14.05514381 16.1433903 18.13576395  9.41031976 13.32428901 13.33349153
#> [22,] 14.62233952 16.7948572 18.86763319  9.79007347 13.86199105 13.87156494
#> [23,] 15.14889320 17.3996438 19.54706085 10.14261618 14.36116442 14.37108306
#> [24,] 15.63639970 17.9595817 20.17610477 10.46901569 14.82332102 14.83355886
#> [25,] 16.08667688 18.4767589 20.75711060 10.77048911 15.25018419 15.26071684
#> [26,] 16.50168602 18.9534281 21.29260907 11.04834957 15.64361323 15.65441761
#> [27,] 16.88346798 19.3919334 21.78523352 11.30396348 16.00554287 16.01659722
#> [28,] 17.23409259 19.7946522 22.23765472 11.53871666 16.33793532 16.34921924
#> [29,] 17.55561940 20.1639499 22.65253019 11.75398804 16.64274303 16.65423746
#> [30,] 17.85006773 20.5021460 23.03246550 11.95112960 16.92188031 16.93356753
#>              [,7]        [,8]        [,9]       [,10]
#>  [1,]  0.03763483  0.03544816  0.04603067  0.06321289
#>  [2,]  0.16327146  0.15378503  0.19969521  0.27423695
#>  [3,]  0.40842606  0.38469562  0.49954064  0.68600795
#>  [4,]  0.78182405  0.73639837  0.95623889  1.31318142
#>  [5,]  1.27777219  1.20353084  1.56282666  2.14619478
#>  [6,]  1.88193532  1.77259078  2.30177077  3.16097015
#>  [7,]  2.57556431  2.42591841  3.15013942  4.32601578
#>  [8,]  3.33839969  3.14443138  4.08315351  5.60730310
#>  [9,]  4.15055422  3.90939797  5.07648923  6.97142875
#> [10,]  4.99364886  4.70350697  6.10766738  8.38752257
#> [11,]  5.85142312  5.51144267  7.15679998  9.82827283
#> [12,]  6.70998855  6.32012358  8.20690027 11.27035200
#> [13,]  7.55784966  7.11872211  9.24390824 12.69445177
#> [14,]  8.38578111  7.89854893 10.25654049 14.08507692
#> [15,]  9.18662372  8.65286084 11.23604074 15.43020264
#> [16,]  9.95504164  9.37663200 12.17588276 16.72086659
#> [17,] 10.68726854 10.06631492 13.07146002 17.95074275
#> [18,] 11.38086069 10.71960785 13.91978361 19.11572650
#> [19,] 12.03446717 11.33523837 14.71920125 20.21354881
#> [20,] 12.64762285 11.91276837 15.46914404 21.24342841
#> [21,] 13.22056628 12.45242254 16.16990375 22.20576599
#> [22,] 13.75408259 12.95494038 16.82244065 23.10188028
#> [23,] 14.24937015 13.42145066 17.42822046 23.93378410
#> [24,] 14.70792910 13.85336633 17.98907799 24.70399716
#> [25,] 15.13146936 14.25229798 18.50710460 25.41539149
#> [26,] 15.52183576 14.61998324 18.98455670 26.07106575
#> [27,] 15.88094797 14.95823024 19.42378220 26.67424426
#> [28,] 16.21075291 15.26887280 19.82716236 27.22819719
#> [29,] 16.51318785 15.55373560 20.19706662 27.73617841
#> [30,] 16.79015219 15.81460771 20.53581813 28.20137826

# multiple stochastic iterations
# with parameter uncertainty
mass(dat) <- list(mu = list(a = 9.63e-6,b = 3.173), cv = list(a = 0.05, b = 0.05))
mass(dat)
#>              [,1]        [,2]        [,3]       [,4]        [,5]        [,6]
#>  [1,]  0.05131621  0.09268822  0.03574002 0.02703816  0.04066922  0.02984621
#>  [2,]  0.23500921  0.46187454  0.15394652 0.10951554  0.17622920  0.12186052
#>  [3,]  0.60810323  1.25987551  0.38338305 0.26244968  0.44051781  0.29349771
#>  [4,]  1.19227208  2.56419303  0.73156772 0.48735607  0.84281868  0.54694387
#>  [5,]  1.98421990  4.38975437  1.19277760 0.77840909  1.37691907  0.87592642
#>  [6,]  2.96444886  6.70608733  1.75344222 1.12587705  2.02733525  1.26960202
#>  [7,]  4.10429362  9.45374800  2.39604645 1.51838666  2.77386040  1.71514989
#>  [8,]  5.37106573 12.55765254  3.10178883 1.94435879  3.59468368  2.19943162
#>  [9,]  6.73157550 15.93697191  3.85229228 2.39287008  4.46841007  2.70999723
#> [10,]  8.15438555 19.51197307  4.63062742 2.85412535  5.37527739  3.23564380
#> [11,]  9.61112662 23.20840286  5.42185524 3.31967215  6.29780891  3.76667591
#> [12,] 11.07714892 26.96001206  6.21324514 3.78244990  7.22108297  4.29497338
#> [13,] 12.53172187 30.70973827  6.99428263 4.23673782  8.13275318  4.81393996
#> [14,] 13.95794158 34.40996587  7.75654813 4.67804514  9.02291489  5.31838364
#> [15,] 15.34246134 38.02218402  8.49352357 5.10297284  9.88388482  5.80436230
#> [16,] 16.67512645 41.51628072  9.20036526 5.50906576 10.70993920  6.26901703
#> [17,] 17.94856882 44.86964268  9.87366823 5.89466714 11.49704058  6.71040695
#> [18,] 19.15779774 48.06617875 10.51123781 6.25878233 12.24257210  7.12735401
#> [19,] 20.29980947 51.09534486 11.11187789 6.60095557 12.94509043  7.51930213
#> [20,] 21.37322893 53.95121930 11.67520049 6.92116137 13.60410316  7.88619271
#> [21,] 22.37798984 56.63165662 12.20145860 7.21971057 14.21987317  8.22835686
#> [22,] 23.31505531 59.13753390 12.69140220 7.49717056 14.79324978  8.54642354
#> [23,] 24.18617838 61.47209347 13.14615627 7.75429844 15.32552558  8.84124267
#> [24,] 24.99369976 63.64037991 13.56711880 7.99198577 15.81831671  9.11382135
#> [25,] 25.74037958 65.64876520 13.95587690 8.21121355 16.27346404  9.36527185
#> [26,] 26.42925927 67.50455414 14.31413848 8.41301594 16.69295288  9.59676971
#> [27,] 27.06354975 69.21566090 14.64367749 8.59845151 17.07884844  9.80952037
#> [28,] 27.64654217 70.79034767 14.94629072 8.76858074 17.43324480 10.00473314
#> [29,] 28.18153790 72.23701660 15.22376416 8.92444883 17.75822529 10.18360117
#> [30,] 28.67179448 73.56404704 15.47784749 9.06707275 18.05583228 10.34728639
#>             [,7]       [,8]        [,9]       [,10]
#>  [1,] 0.02082147  0.0621021  0.04499299  0.04248497
#>  [2,] 0.08084208  0.2885579  0.19773755  0.18763807
#>  [3,] 0.18868116  0.7534585  0.49866198  0.47465269
#>  [4,] 0.34387423  1.4867685  0.96004211  0.91581434
#>  [5,] 0.54151555  2.4863684  1.57585891  1.50574452
#>  [6,] 0.77454490  3.7289011  2.32890854  2.22818555
#>  [7,] 1.03516538  5.1787097  3.19611666  3.06111314
#>  [8,] 1.31569698  6.7944870  4.15224306  3.98032071
#>  [9,] 1.60906069  8.5338930  5.17231582  4.96178642
#> [10,] 1.90902295 10.3565509  6.23312234  5.98313108
#> [11,] 2.21028719 12.2258294  7.31402866  7.02442577
#> [12,] 2.50849129 14.1097544  8.39733805  8.06855167
#> [13,] 2.80014998 15.9813233  9.46834604  9.10126362
#> [14,] 3.08256812 17.8184274 10.51520579 10.11106796
#> [15,] 3.35374161 19.6035343 11.52868412 11.08899236
#> [16,] 3.61225654 21.3232372 12.50186346 12.02830166
#> [17,] 3.85719277 22.9677443 13.42982624 12.92419543
#> [18,] 4.08803560 24.5303583 14.30934517 13.77351018
#> [19,] 4.30459712 26.0069763 15.13859348 14.57444023
#> [20,] 4.50694769 27.3956276 15.91688272 15.32628457
#> [21,] 4.69535738 28.6960604 16.64443126 16.02922319
#> [22,] 4.87024661 29.9093792 17.32216405 16.68412330
#> [23,] 5.03214521 31.0377326 17.95154229 17.29237443
#> [24,] 5.18165873 32.0840499 18.53442058 17.85574995
#> [25,] 5.31944128 33.0518208 19.07292892 18.37629260
#> [26,] 5.44617376 33.9449145 19.56937642 18.85622105
#> [27,] 5.56254675 34.7674326 20.02617385 19.29785462
#> [28,] 5.66924721 35.5235920 20.44577226 19.70355360
#> [29,] 5.76694850 36.2176319 20.83061510 20.07567272
#> [30,] 5.85630285 36.8537429 21.18310168 20.41652543