Access or assign maturity at age

This function can be used to access or assign the maturity at age in an lhm object.

Assignment can be of a list or numeric vector. When a list is assigned it should provide the \(a_{crit}\) and optionally the \(\delta\) parameter. If only acrit is supplied then maturity is knife-edged at that age. If both acrit and delta are supplied then maturity-at-age is calculated as a logistic ogive: $$Maturity = 1/(1+exp((acrit-Age)/delta)).$$

Uncertainty accross iterations in the lhm object can be included by providing a cv for each of the parameters. Parameter uncertainty is assumed to be log-normal with \(\sigma = \sqrt{ln(1+cv^2)}\).

When assigning a numeric vector it is assumed that each replicate is identical.

Usage

maturity(object, ...)

# S4 method for class 'lhm'
maturity(object)

maturity(object) <- value

# S4 method for class 'lhm,list'
maturity(object) <- value

# S4 method for class 'lhm,numeric'
maturity(object) <- value

Arguments

object

a lhm object

...

arguments to generic function

value

either a named list or numeric vector

numeric

a vector of length equal to the maximum age

list

a list object containing the components mu and cv

mu

a list containing the parameters acrit and (optionally) delta representing their mean values

cv

optional single numeric value or a list containing parameters acrit and delta representing their coefficients of variation

Value

Accessor function returns a matrix of maturity at age across iterations. Assignment function populates the matrix across iterations.

Examples

# initialize single iteration
dat <- lhm(ainf = 30, iter = 1)

# add knife-edged maturity
maturity(dat) <- list(mu = list(acrit = 10))
maturity(dat)
#>       [,1]
#>  [1,]    0
#>  [2,]    0
#>  [3,]    0
#>  [4,]    0
#>  [5,]    0
#>  [6,]    0
#>  [7,]    0
#>  [8,]    0
#>  [9,]    0
#> [10,]    1
#> [11,]    1
#> [12,]    1
#> [13,]    1
#> [14,]    1
#> [15,]    1
#> [16,]    1
#> [17,]    1
#> [18,]    1
#> [19,]    1
#> [20,]    1
#> [21,]    1
#> [22,]    1
#> [23,]    1
#> [24,]    1
#> [25,]    1
#> [26,]    1
#> [27,]    1
#> [28,]    1
#> [29,]    1
#> [30,]    1

# increase iterations
dat <- iterate(dat, iter = 10)
maturity(dat)
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#>  [1,]    0    0    0    0    0    0    0    0    0     0
#>  [2,]    0    0    0    0    0    0    0    0    0     0
#>  [3,]    0    0    0    0    0    0    0    0    0     0
#>  [4,]    0    0    0    0    0    0    0    0    0     0
#>  [5,]    0    0    0    0    0    0    0    0    0     0
#>  [6,]    0    0    0    0    0    0    0    0    0     0
#>  [7,]    0    0    0    0    0    0    0    0    0     0
#>  [8,]    0    0    0    0    0    0    0    0    0     0
#>  [9,]    0    0    0    0    0    0    0    0    0     0
#> [10,]    1    1    1    1    1    1    1    1    1     1
#> [11,]    1    1    1    1    1    1    1    1    1     1
#> [12,]    1    1    1    1    1    1    1    1    1     1
#> [13,]    1    1    1    1    1    1    1    1    1     1
#> [14,]    1    1    1    1    1    1    1    1    1     1
#> [15,]    1    1    1    1    1    1    1    1    1     1
#> [16,]    1    1    1    1    1    1    1    1    1     1
#> [17,]    1    1    1    1    1    1    1    1    1     1
#> [18,]    1    1    1    1    1    1    1    1    1     1
#> [19,]    1    1    1    1    1    1    1    1    1     1
#> [20,]    1    1    1    1    1    1    1    1    1     1
#> [21,]    1    1    1    1    1    1    1    1    1     1
#> [22,]    1    1    1    1    1    1    1    1    1     1
#> [23,]    1    1    1    1    1    1    1    1    1     1
#> [24,]    1    1    1    1    1    1    1    1    1     1
#> [25,]    1    1    1    1    1    1    1    1    1     1
#> [26,]    1    1    1    1    1    1    1    1    1     1
#> [27,]    1    1    1    1    1    1    1    1    1     1
#> [28,]    1    1    1    1    1    1    1    1    1     1
#> [29,]    1    1    1    1    1    1    1    1    1     1
#> [30,]    1    1    1    1    1    1    1    1    1     1

# add logistic maturity ogive
# with uncertainty
maturity(dat) <- list(mu=list(acrit = 1, delta = 3), 
                         cv=list(acrit = 0.01, delta = 0.01))
maturity(dat)
#>            [,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#>  [1,] 0.4999111 0.5001126 0.5015800 0.5002352 0.5002646 0.4990928 0.4999463
#>  [2,] 0.5821280 0.5804792 0.5840727 0.5823360 0.5827359 0.5832563 0.5818260
#>  [3,] 0.6600205 0.6567921 0.6621097 0.6601136 0.6608246 0.6628317 0.6594316
#>  [4,] 0.7301249 0.7257811 0.7322180 0.7301197 0.7310443 0.7341386 0.7293349
#>  [5,] 0.7903650 0.7854328 0.7923419 0.7902876 0.7913172 0.7950285 0.7894710
#>  [6,] 0.8401034 0.8350574 0.8418815 0.8399814 0.8410204 0.8449187 0.8391920
#>  [7,] 0.8798350 0.8750309 0.8813723 0.8796923 0.8806706 0.8844308 0.8789705
#>  [8,] 0.9107428 0.9064028 0.9120306 0.9105973 0.9114728 0.9148899 0.9099648
#>  [9,] 0.9342945 0.9305245 0.9353468 0.9341575 0.9349118 0.9378851 0.9336213
#> [10,] 0.9519596 0.9487808 0.9528028 0.9518373 0.9524685 0.9549732 0.9513941
#> [11,] 0.9650530 0.9624336 0.9657186 0.9649479 0.9654645 0.9675231 0.9645887
#> [12,] 0.9746728 0.9725524 0.9751919 0.9745849 0.9750004 0.9766605 0.9742983
#> [13,] 0.9816947 0.9800023 0.9820959 0.9816228 0.9819525 0.9832716 0.9813969
#> [14,] 0.9867962 0.9854603 0.9871039 0.9867383 0.9869970 0.9880330 0.9865620
#> [15,] 0.9904898 0.9894447 0.9907244 0.9904438 0.9906450 0.9914509 0.9903071
#> [16,] 0.9931573 0.9923457 0.9933353 0.9931211 0.9932765 0.9938987 0.9930159
#> [17,] 0.9950803 0.9944539 0.9952148 0.9950520 0.9951714 0.9956487 0.9949716
#> [18,] 0.9964648 0.9959838 0.9965662 0.9964429 0.9965341 0.9968983 0.9963816
#> [19,] 0.9974607 0.9970929 0.9975368 0.9974438 0.9975132 0.9977898 0.9973973
#> [20,] 0.9981765 0.9978963 0.9982336 0.9981636 0.9982162 0.9984255 0.9981284
#> [21,] 0.9986908 0.9984781 0.9987335 0.9986810 0.9987207 0.9988786 0.9986544
#> [22,] 0.9990602 0.9988991 0.9990921 0.9990527 0.9990827 0.9992014 0.9990327
#> [23,] 0.9993255 0.9992038 0.9993492 0.9993198 0.9993423 0.9994313 0.9993047
#> [24,] 0.9995159 0.9994242 0.9995335 0.9995116 0.9995285 0.9995951 0.9995003
#> [25,] 0.9996525 0.9995836 0.9996657 0.9996493 0.9996620 0.9997117 0.9996409
#> [26,] 0.9997507 0.9996989 0.9997604 0.9997482 0.9997577 0.9997947 0.9997419
#> [27,] 0.9998211 0.9997823 0.9998283 0.9998192 0.9998263 0.9998539 0.9998145
#> [28,] 0.9998716 0.9998426 0.9998769 0.9998702 0.9998755 0.9998960 0.9998667
#> [29,] 0.9999079 0.9998862 0.9999118 0.9999068 0.9999107 0.9999259 0.9999042
#> [30,] 0.9999339 0.9999177 0.9999368 0.9999331 0.9999360 0.9999473 0.9999312
#>            [,8]      [,9]     [,10]
#>  [1,] 0.5005364 0.5000459 0.5002134
#>  [2,] 0.5832932 0.5827861 0.5827663
#>  [3,] 0.6616082 0.6611132 0.6609265
#>  [4,] 0.7319681 0.7315099 0.7311985
#>  [5,] 0.7922913 0.7918865 0.7915012
#>  [6,] 0.8419693 0.8416254 0.8412134
#>  [7,] 0.8815422 0.8812593 0.8808569
#>  [8,] 0.9122385 0.9120116 0.9116425
#>  [9,] 0.9355618 0.9353834 0.9350597
#> [10,] 0.9530060 0.9528679 0.9525935
#> [11,] 0.9659000 0.9657944 0.9655675
#> [12,] 0.9753477 0.9752676 0.9750837
#> [13,] 0.9822260 0.9821657 0.9820189
#> [14,] 0.9872104 0.9871652 0.9870493
#> [15,] 0.9908100 0.9907763 0.9906858
#> [16,] 0.9934033 0.9933782 0.9933081
#> [17,] 0.9952683 0.9952497 0.9951957
#> [18,] 0.9966079 0.9965941 0.9965527
#> [19,] 0.9975691 0.9975589 0.9975274
#> [20,] 0.9982584 0.9982509 0.9982270
#> [21,] 0.9987525 0.9987470 0.9987289
#> [22,] 0.9991066 0.9991025 0.9990888
#> [23,] 0.9993602 0.9993572 0.9993469
#> [24,] 0.9995419 0.9995396 0.9995319
#> [25,] 0.9996720 0.9996703 0.9996646
#> [26,] 0.9997651 0.9997639 0.9997596
#> [27,] 0.9998318 0.9998310 0.9998277
#> [28,] 0.9998796 0.9998790 0.9998766
#> [29,] 0.9999138 0.9999133 0.9999115
#> [30,] 0.9999383 0.9999379 0.9999366