Access or assign natural mortality at age

This function can be used to access or assign the natural mortality at age in an lhm object.

Usage

nmort(object, ...)

# S4 method for class 'lhm'
nmort(object)

nmort(object) <- value

# S4 method for class 'lhm,list'
nmort(object) <- value

# S4 method for class 'lhm,numeric'
nmort(object) <- value

Arguments

object

a lhm object

value

either a named list or numeric vector

numeric

a single numeric or vector of values

list

a list object containing the components mu, cv and range

mu

a single numeric or vector of values representing the mean natural mortality

cv

optional single numeric value representing the coefficient of variation

range

optional vector of length two containing the lower and upper bounds

Value

Accessor function returns a matrix of maturity at age across iterations. Assignment function populates the matrix across iterations.

Details

Assignment can be of a list or numeric vector. When a list is assigned it should contain a value for mu, which is the mean natural mortality at age and may or may not be constant across ages.

If the list also contains cv then log-normal uncertainty is added around the mortality at age. If the list contains range then the uncertainty is bounded at these values. If range is provided but not cv, and if mortality is constant across ages, then a uniform uncertainy is implemented.

Examples

# M at age vector input
dat <- lhm(ainf = 10,iter=1)
nmort(dat) <- c(0.1,0.1,0.2)
nmort(dat)
#>       [,1]
#>  [1,]  0.1
#>  [2,]  0.1
#>  [3,]  0.2
#>  [4,]  0.2
#>  [5,]  0.2
#>  [6,]  0.2
#>  [7,]  0.2
#>  [8,]  0.2
#>  [9,]  0.2
#> [10,]  0.2

# constant M with log-normal uncertainty
dat <- lhm(ainf=10, iter = 20)
nmort(dat) <- list(mu = 0.2, cv = 0.25)
nmort(dat)
#>            [,1]      [,2]     [,3]      [,4]      [,5]     [,6]      [,7]
#>  [1,] 0.2221569 0.1108381 0.375593 0.1775167 0.2044938 0.252783 0.1647199
#>  [2,] 0.2221569 0.1108381 0.375593 0.1775167 0.2044938 0.252783 0.1647199
#>  [3,] 0.2221569 0.1108381 0.375593 0.1775167 0.2044938 0.252783 0.1647199
#>  [4,] 0.2221569 0.1108381 0.375593 0.1775167 0.2044938 0.252783 0.1647199
#>  [5,] 0.2221569 0.1108381 0.375593 0.1775167 0.2044938 0.252783 0.1647199
#>  [6,] 0.2221569 0.1108381 0.375593 0.1775167 0.2044938 0.252783 0.1647199
#>  [7,] 0.2221569 0.1108381 0.375593 0.1775167 0.2044938 0.252783 0.1647199
#>  [8,] 0.2221569 0.1108381 0.375593 0.1775167 0.2044938 0.252783 0.1647199
#>  [9,] 0.2221569 0.1108381 0.375593 0.1775167 0.2044938 0.252783 0.1647199
#> [10,] 0.2221569 0.1108381 0.375593 0.1775167 0.2044938 0.252783 0.1647199
#>            [,8]      [,9]     [,10]     [,11]     [,12]    [,13]     [,14]
#>  [1,] 0.2552601 0.1826296 0.1451931 0.1525863 0.2522055 0.200422 0.2188345
#>  [2,] 0.2552601 0.1826296 0.1451931 0.1525863 0.2522055 0.200422 0.2188345
#>  [3,] 0.2552601 0.1826296 0.1451931 0.1525863 0.2522055 0.200422 0.2188345
#>  [4,] 0.2552601 0.1826296 0.1451931 0.1525863 0.2522055 0.200422 0.2188345
#>  [5,] 0.2552601 0.1826296 0.1451931 0.1525863 0.2522055 0.200422 0.2188345
#>  [6,] 0.2552601 0.1826296 0.1451931 0.1525863 0.2522055 0.200422 0.2188345
#>  [7,] 0.2552601 0.1826296 0.1451931 0.1525863 0.2522055 0.200422 0.2188345
#>  [8,] 0.2552601 0.1826296 0.1451931 0.1525863 0.2522055 0.200422 0.2188345
#>  [9,] 0.2552601 0.1826296 0.1451931 0.1525863 0.2522055 0.200422 0.2188345
#> [10,] 0.2552601 0.1826296 0.1451931 0.1525863 0.2522055 0.200422 0.2188345
#>           [,15]     [,16]     [,17]     [,18]     [,19]     [,20]
#>  [1,] 0.1276849 0.1350816 0.2080896 0.2696921 0.2056724 0.2684316
#>  [2,] 0.1276849 0.1350816 0.2080896 0.2696921 0.2056724 0.2684316
#>  [3,] 0.1276849 0.1350816 0.2080896 0.2696921 0.2056724 0.2684316
#>  [4,] 0.1276849 0.1350816 0.2080896 0.2696921 0.2056724 0.2684316
#>  [5,] 0.1276849 0.1350816 0.2080896 0.2696921 0.2056724 0.2684316
#>  [6,] 0.1276849 0.1350816 0.2080896 0.2696921 0.2056724 0.2684316
#>  [7,] 0.1276849 0.1350816 0.2080896 0.2696921 0.2056724 0.2684316
#>  [8,] 0.1276849 0.1350816 0.2080896 0.2696921 0.2056724 0.2684316
#>  [9,] 0.1276849 0.1350816 0.2080896 0.2696921 0.2056724 0.2684316
#> [10,] 0.1276849 0.1350816 0.2080896 0.2696921 0.2056724 0.2684316

# constant M with bounded log-normal uncertainty
dat <- lhm(ainf = 10, iter = 20)
nmort(dat) <- list(mu = 0.2, cv = 0.5, range = c(0.1,0.3))
nmort(dat)
#>            [,1]      [,2]      [,3]      [,4]      [,5] [,6] [,7]      [,8]
#>  [1,] 0.2291165 0.2382598 0.1698327 0.1940432 0.1714131  0.3  0.3 0.2547952
#>  [2,] 0.2291165 0.2382598 0.1698327 0.1940432 0.1714131  0.3  0.3 0.2547952
#>  [3,] 0.2291165 0.2382598 0.1698327 0.1940432 0.1714131  0.3  0.3 0.2547952
#>  [4,] 0.2291165 0.2382598 0.1698327 0.1940432 0.1714131  0.3  0.3 0.2547952
#>  [5,] 0.2291165 0.2382598 0.1698327 0.1940432 0.1714131  0.3  0.3 0.2547952
#>  [6,] 0.2291165 0.2382598 0.1698327 0.1940432 0.1714131  0.3  0.3 0.2547952
#>  [7,] 0.2291165 0.2382598 0.1698327 0.1940432 0.1714131  0.3  0.3 0.2547952
#>  [8,] 0.2291165 0.2382598 0.1698327 0.1940432 0.1714131  0.3  0.3 0.2547952
#>  [9,] 0.2291165 0.2382598 0.1698327 0.1940432 0.1714131  0.3  0.3 0.2547952
#> [10,] 0.2291165 0.2382598 0.1698327 0.1940432 0.1714131  0.3  0.3 0.2547952
#>            [,9]     [,10] [,11] [,12]     [,13]     [,14]    [,15]     [,16]
#>  [1,] 0.2326407 0.1380698   0.3   0.1 0.1662012 0.2195778 0.149354 0.2185738
#>  [2,] 0.2326407 0.1380698   0.3   0.1 0.1662012 0.2195778 0.149354 0.2185738
#>  [3,] 0.2326407 0.1380698   0.3   0.1 0.1662012 0.2195778 0.149354 0.2185738
#>  [4,] 0.2326407 0.1380698   0.3   0.1 0.1662012 0.2195778 0.149354 0.2185738
#>  [5,] 0.2326407 0.1380698   0.3   0.1 0.1662012 0.2195778 0.149354 0.2185738
#>  [6,] 0.2326407 0.1380698   0.3   0.1 0.1662012 0.2195778 0.149354 0.2185738
#>  [7,] 0.2326407 0.1380698   0.3   0.1 0.1662012 0.2195778 0.149354 0.2185738
#>  [8,] 0.2326407 0.1380698   0.3   0.1 0.1662012 0.2195778 0.149354 0.2185738
#>  [9,] 0.2326407 0.1380698   0.3   0.1 0.1662012 0.2195778 0.149354 0.2185738
#> [10,] 0.2326407 0.1380698   0.3   0.1 0.1662012 0.2195778 0.149354 0.2185738
#>           [,17]     [,18]     [,19]     [,20]
#>  [1,] 0.2955785 0.2935762 0.1756945 0.2250658
#>  [2,] 0.2955785 0.2935762 0.1756945 0.2250658
#>  [3,] 0.2955785 0.2935762 0.1756945 0.2250658
#>  [4,] 0.2955785 0.2935762 0.1756945 0.2250658
#>  [5,] 0.2955785 0.2935762 0.1756945 0.2250658
#>  [6,] 0.2955785 0.2935762 0.1756945 0.2250658
#>  [7,] 0.2955785 0.2935762 0.1756945 0.2250658
#>  [8,] 0.2955785 0.2935762 0.1756945 0.2250658
#>  [9,] 0.2955785 0.2935762 0.1756945 0.2250658
#> [10,] 0.2955785 0.2935762 0.1756945 0.2250658

# constant M with uniform uncertainty
dat <- lhm(ainf = 10,iter = 20)
nmort(dat) <- list(range = c(0.1,0.3))
nmort(dat)
#>           [,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#>  [1,] 0.290565 0.1643709 0.1723068 0.2775447 0.2656029 0.1201313 0.2812103
#>  [2,] 0.290565 0.1643709 0.1723068 0.2775447 0.2656029 0.1201313 0.2812103
#>  [3,] 0.290565 0.1643709 0.1723068 0.2775447 0.2656029 0.1201313 0.2812103
#>  [4,] 0.290565 0.1643709 0.1723068 0.2775447 0.2656029 0.1201313 0.2812103
#>  [5,] 0.290565 0.1643709 0.1723068 0.2775447 0.2656029 0.1201313 0.2812103
#>  [6,] 0.290565 0.1643709 0.1723068 0.2775447 0.2656029 0.1201313 0.2812103
#>  [7,] 0.290565 0.1643709 0.1723068 0.2775447 0.2656029 0.1201313 0.2812103
#>  [8,] 0.290565 0.1643709 0.1723068 0.2775447 0.2656029 0.1201313 0.2812103
#>  [9,] 0.290565 0.1643709 0.1723068 0.2775447 0.2656029 0.1201313 0.2812103
#> [10,] 0.290565 0.1643709 0.1723068 0.2775447 0.2656029 0.1201313 0.2812103
#>            [,8]      [,9]     [,10]     [,11]     [,12]   [,13]     [,14]
#>  [1,] 0.2545461 0.1766741 0.2999305 0.1698598 0.2894637 0.14322 0.1064185
#>  [2,] 0.2545461 0.1766741 0.2999305 0.1698598 0.2894637 0.14322 0.1064185
#>  [3,] 0.2545461 0.1766741 0.2999305 0.1698598 0.2894637 0.14322 0.1064185
#>  [4,] 0.2545461 0.1766741 0.2999305 0.1698598 0.2894637 0.14322 0.1064185
#>  [5,] 0.2545461 0.1766741 0.2999305 0.1698598 0.2894637 0.14322 0.1064185
#>  [6,] 0.2545461 0.1766741 0.2999305 0.1698598 0.2894637 0.14322 0.1064185
#>  [7,] 0.2545461 0.1766741 0.2999305 0.1698598 0.2894637 0.14322 0.1064185
#>  [8,] 0.2545461 0.1766741 0.2999305 0.1698598 0.2894637 0.14322 0.1064185
#>  [9,] 0.2545461 0.1766741 0.2999305 0.1698598 0.2894637 0.14322 0.1064185
#> [10,] 0.2545461 0.1766741 0.2999305 0.1698598 0.2894637 0.14322 0.1064185
#>           [,15]     [,16]     [,17]     [,18]     [,19]    [,20]
#>  [1,] 0.1290632 0.2708768 0.1426299 0.1420621 0.1079041 0.288955
#>  [2,] 0.1290632 0.2708768 0.1426299 0.1420621 0.1079041 0.288955
#>  [3,] 0.1290632 0.2708768 0.1426299 0.1420621 0.1079041 0.288955
#>  [4,] 0.1290632 0.2708768 0.1426299 0.1420621 0.1079041 0.288955
#>  [5,] 0.1290632 0.2708768 0.1426299 0.1420621 0.1079041 0.288955
#>  [6,] 0.1290632 0.2708768 0.1426299 0.1420621 0.1079041 0.288955
#>  [7,] 0.1290632 0.2708768 0.1426299 0.1420621 0.1079041 0.288955
#>  [8,] 0.1290632 0.2708768 0.1426299 0.1420621 0.1079041 0.288955
#>  [9,] 0.1290632 0.2708768 0.1426299 0.1420621 0.1079041 0.288955
#> [10,] 0.1290632 0.2708768 0.1426299 0.1420621 0.1079041 0.288955

# M at age with log-normal uncertainty
dat <- lhm(ainf=10, iter = 20)
nmort(dat) <- list(mu = c(0.1,0.1,0.2), cv = 0.25)
nmort(dat)
#>             [,1]       [,2]       [,3]      [,4]      [,5]       [,6]
#>  [1,] 0.08184549 0.08454917 0.08500587 0.1025942 0.1234423 0.09215083
#>  [2,] 0.08184549 0.08454917 0.08500587 0.1025942 0.1234423 0.09215083
#>  [3,] 0.16369097 0.16909834 0.17001173 0.2051884 0.2468846 0.18430165
#>  [4,] 0.16369097 0.16909834 0.17001173 0.2051884 0.2468846 0.18430165
#>  [5,] 0.16369097 0.16909834 0.17001173 0.2051884 0.2468846 0.18430165
#>  [6,] 0.16369097 0.16909834 0.17001173 0.2051884 0.2468846 0.18430165
#>  [7,] 0.16369097 0.16909834 0.17001173 0.2051884 0.2468846 0.18430165
#>  [8,] 0.16369097 0.16909834 0.17001173 0.2051884 0.2468846 0.18430165
#>  [9,] 0.16369097 0.16909834 0.17001173 0.2051884 0.2468846 0.18430165
#> [10,] 0.16369097 0.16909834 0.17001173 0.2051884 0.2468846 0.18430165
#>             [,7]      [,8]       [,9]     [,10]     [,11]     [,12]      [,13]
#>  [1,] 0.06873744 0.1033907 0.08701476 0.1115931 0.1637814 0.1077126 0.06408132
#>  [2,] 0.06873744 0.1033907 0.08701476 0.1115931 0.1637814 0.1077126 0.06408132
#>  [3,] 0.13747489 0.2067814 0.17402951 0.2231861 0.3275627 0.2154253 0.12816264
#>  [4,] 0.13747489 0.2067814 0.17402951 0.2231861 0.3275627 0.2154253 0.12816264
#>  [5,] 0.13747489 0.2067814 0.17402951 0.2231861 0.3275627 0.2154253 0.12816264
#>  [6,] 0.13747489 0.2067814 0.17402951 0.2231861 0.3275627 0.2154253 0.12816264
#>  [7,] 0.13747489 0.2067814 0.17402951 0.2231861 0.3275627 0.2154253 0.12816264
#>  [8,] 0.13747489 0.2067814 0.17402951 0.2231861 0.3275627 0.2154253 0.12816264
#>  [9,] 0.13747489 0.2067814 0.17402951 0.2231861 0.3275627 0.2154253 0.12816264
#> [10,] 0.13747489 0.2067814 0.17402951 0.2231861 0.3275627 0.2154253 0.12816264
#>           [,14]     [,15]     [,16]     [,17]     [,18]     [,19]      [,20]
#>  [1,] 0.1031584 0.1263446 0.1602913 0.1083669 0.1366675 0.1077579 0.09961844
#>  [2,] 0.1031584 0.1263446 0.1602913 0.1083669 0.1366675 0.1077579 0.09961844
#>  [3,] 0.2063168 0.2526891 0.3205826 0.2167338 0.2733351 0.2155159 0.19923688
#>  [4,] 0.2063168 0.2526891 0.3205826 0.2167338 0.2733351 0.2155159 0.19923688
#>  [5,] 0.2063168 0.2526891 0.3205826 0.2167338 0.2733351 0.2155159 0.19923688
#>  [6,] 0.2063168 0.2526891 0.3205826 0.2167338 0.2733351 0.2155159 0.19923688
#>  [7,] 0.2063168 0.2526891 0.3205826 0.2167338 0.2733351 0.2155159 0.19923688
#>  [8,] 0.2063168 0.2526891 0.3205826 0.2167338 0.2733351 0.2155159 0.19923688
#>  [9,] 0.2063168 0.2526891 0.3205826 0.2167338 0.2733351 0.2155159 0.19923688
#> [10,] 0.2063168 0.2526891 0.3205826 0.2167338 0.2733351 0.2155159 0.19923688

# M at age with bounded log-normal uncertainty
dat <- lhm(ainf = 10, iter = 20)
nmort(dat) <- list(mu = c(0.1,0.1,0.2), cv = 0.25, range = c(0.1,0.3))
nmort(dat)
#>            [,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#>  [1,] 0.1000000 0.1125592 0.1000000 0.1000000 0.1017144 0.1000000 0.1194868
#>  [2,] 0.1000000 0.1125592 0.1000000 0.1000000 0.1017144 0.1000000 0.1194868
#>  [3,] 0.1950965 0.2251184 0.1818778 0.1703637 0.2034288 0.1463187 0.2389736
#>  [4,] 0.1950965 0.2251184 0.1818778 0.1703637 0.2034288 0.1463187 0.2389736
#>  [5,] 0.1950965 0.2251184 0.1818778 0.1703637 0.2034288 0.1463187 0.2389736
#>  [6,] 0.1950965 0.2251184 0.1818778 0.1703637 0.2034288 0.1463187 0.2389736
#>  [7,] 0.1950965 0.2251184 0.1818778 0.1703637 0.2034288 0.1463187 0.2389736
#>  [8,] 0.1950965 0.2251184 0.1818778 0.1703637 0.2034288 0.1463187 0.2389736
#>  [9,] 0.1950965 0.2251184 0.1818778 0.1703637 0.2034288 0.1463187 0.2389736
#> [10,] 0.1950965 0.2251184 0.1818778 0.1703637 0.2034288 0.1463187 0.2389736
#>            [,8]     [,9]     [,10]     [,11]     [,12]     [,13]     [,14]
#>  [1,] 0.1000000 0.100000 0.1000000 0.1084974 0.1000000 0.1000000 0.1209305
#>  [2,] 0.1000000 0.100000 0.1000000 0.1084974 0.1000000 0.1000000 0.1209305
#>  [3,] 0.1979721 0.140704 0.1537531 0.2169948 0.1571871 0.1807953 0.2418610
#>  [4,] 0.1979721 0.140704 0.1537531 0.2169948 0.1571871 0.1807953 0.2418610
#>  [5,] 0.1979721 0.140704 0.1537531 0.2169948 0.1571871 0.1807953 0.2418610
#>  [6,] 0.1979721 0.140704 0.1537531 0.2169948 0.1571871 0.1807953 0.2418610
#>  [7,] 0.1979721 0.140704 0.1537531 0.2169948 0.1571871 0.1807953 0.2418610
#>  [8,] 0.1979721 0.140704 0.1537531 0.2169948 0.1571871 0.1807953 0.2418610
#>  [9,] 0.1979721 0.140704 0.1537531 0.2169948 0.1571871 0.1807953 0.2418610
#> [10,] 0.1979721 0.140704 0.1537531 0.2169948 0.1571871 0.1807953 0.2418610
#>           [,15]     [,16]     [,17]     [,18]     [,19]     [,20]
#>  [1,] 0.1000000 0.1000000 0.1000000 0.1539907 0.1171636 0.1234879
#>  [2,] 0.1000000 0.1000000 0.1000000 0.1539907 0.1171636 0.1234879
#>  [3,] 0.1972707 0.1864104 0.1582703 0.3000000 0.2343272 0.2469759
#>  [4,] 0.1972707 0.1864104 0.1582703 0.3000000 0.2343272 0.2469759
#>  [5,] 0.1972707 0.1864104 0.1582703 0.3000000 0.2343272 0.2469759
#>  [6,] 0.1972707 0.1864104 0.1582703 0.3000000 0.2343272 0.2469759
#>  [7,] 0.1972707 0.1864104 0.1582703 0.3000000 0.2343272 0.2469759
#>  [8,] 0.1972707 0.1864104 0.1582703 0.3000000 0.2343272 0.2469759
#>  [9,] 0.1972707 0.1864104 0.1582703 0.3000000 0.2343272 0.2469759
#> [10,] 0.1972707 0.1864104 0.1582703 0.3000000 0.2343272 0.2469759